% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_pct_diff.r
\name{absolute_pct_diff}
\alias{absolute_pct_diff}
\title{Generate the either the average or median absolute percentage difference between two vectors}
\usage{
absolute_pct_diff(x1 = NULL, x2 = NULL, use_median = FALSE)
}
\arguments{
\item{x1}{numeric vector; This is a required entry.}

\item{x2}{numeric vector; This is a required entry, and must be the same length as \code{x1}.}

\item{use_median}{logical scalar, if \code{TRUE}, returns the median of the absolute percentage difference;
if \code{FALSE}, returns the average of the absolute percentage difference
Default is \code{FALSE}.}
}
\value{
Either the average or median absolute percentage difference of the two series.
}
\description{
Generate the average absolute percentage difference (AAPD) or median absolute percentage difference (MAPD)
between two vectors - the vectors must be the same length.
}
\details{
Version 2.0, 6/21/2024
}
\examples{
ukgas_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  x11="", transform.function = "log", forecast.maxlead=20,
                  check.print = c( "pacf", "pacfplot" ))
ukgas_seats_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log",, forecast.maxlead=20,
                  check.print = c( "pacf", "pacfplot" ))
ukgas_x11_sa     <- seasonal::final(ukgas_seas)
ukgas_seats_sa   <- seasonal::final(ukgas_seats_seas)
ukgas_sa_aapd    <- 
   sautilities::absolute_pct_diff(ukgas_seats_sa, ukgas_x11_sa)
ukgas_sa_mapd    <- 
   sautilities::absolute_pct_diff(ukgas_seats_sa, ukgas_x11_sa, use_median = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
