% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_fail_why.r
\name{acf_fail_why}
\alias{acf_fail_why}
\title{ACF Test Explanation}
\usage{
acf_fail_why(
  udg_list = NULL,
  acf_lags_fail = c(1, 2, 3, 4, 12, 24),
  num_sig = 8,
  include_pacf = TRUE,
  return_both = FALSE
)
}
\arguments{
\item{udg_list}{List object generated by \code{udg()} function of the \code{seasonal} package.
This is a required entry.}

\item{acf_lags_fail}{\itemize{
\item lags of the ACF to test
Default is \code{c(1, 2, 3, 4, 12, 24)}.
}}

\item{num_sig}{\itemize{
\item limit for number of lags with significant ACF values
Default is \code{8}.
}}

\item{include_pacf}{Logical scalar that indicates if the PACF is included in the testing.
Default is TRUE}

\item{return_both}{Logical scalar indicating whether the calling function will return both the
test results and why the test failed or just produce a warning. Default is \code{FALSE}.}
}
\value{
Character object that tells why series fails the ACF test, \code{'pass'} otherwise.
}
\description{
Generates text on why the sample autocorrelation of the residuals from a time series model
fails the Ljung-Box or Box-Pierce Q test.
}
\details{
Version 4.1, 2024-05-15
}
\examples{
ukgas_seas <- seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                             x11="", transform.function = "log", forecast.maxlead=20,
                             check.print = c( "pacf", "pacfplot" ))
ukgas_udg <- seasonal::udg(ukgas_seas)
ukgas_acf_fail_why <- acf_fail_why(ukgas_udg, acf_lags_fail = c(1, 2, 3, 4, 8), 
                                   num_sig = 4, return_both = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
