% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_test.r
\name{acf_test}
\alias{acf_test}
\title{Global ACF test}
\usage{
acf_test(
  seas_obj = NULL,
  num_sig = 8,
  acf_lags_fail = c(1, 2, 3, 4, 12, 24),
  acf_lags_warn = c(12, 24),
  include_pacf = TRUE,
  return_this = "test"
)
}
\arguments{
\item{seas_obj}{Object generated by \code{seas()} of the \code{seasonal} package.
This is a required entry.}

\item{num_sig}{Limit for number of lags with significant ACF values
Default is \code{8}.}

\item{acf_lags_fail}{\itemize{
\item lags of the ACF to test
Default is \code{c(1, 2, 3, 4, 12, 24)}.
}}

\item{acf_lags_warn}{\itemize{
\item lags of the ACF to test for warnings
Default is \code{c(12, 24)}.
}}

\item{include_pacf}{Logical scalar that indicates if the PACF is included in the testing.
Default is TRUE}

\item{return_this}{character string; what the function returns -
\code{'test'} returns test results, \code{'why'} returns why the test failed or
received a warning, or \code{'both'}. Default is \code{'test'}.}
}
\value{
A text string denoting if series passes, fails, or has a warning for residual
autocorrelation. If model diagnostics not found, return \code{'none'}.
}
\description{
Tests whether the residuals from a time series model has acceptable autocorrelation in the residuals.
}
\details{
Version 4.4, 5/15/2024
}
\examples{
ukgas_seas <- seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                             x11="", transform.function = "log", forecast.maxlead=20,
                             check.print = c( "pacf", "pacfplot" ))
ukgas_acf_test <- acf_test(ukgas_seas, num_sig = 4, acf_lags_fail = c(1, 2, 3, 4, 8),
                           acf_lags_warn = c(4, 8), return_this = "both")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
