% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_warn.r
\name{acf_warn}
\alias{acf_warn}
\title{ACF test warning message}
\usage{
acf_warn(udg_list = NULL, acf_lags_warn = c(12, 24))
}
\arguments{
\item{udg_list}{\itemize{
\item list object generated by \code{udg()} function of the \code{seasonal} package.
}}

\item{acf_lags_warn}{\itemize{
\item lags of the ACF to test for warnings
}}
}
\value{
Logical object which is \code{TRUE} if series generates a warning for the ACF test,
\code{FALSE} otherwise.
}
\description{
Tests whether the residuals from a time series model generates a warning for the ACF test.
}
\details{
Version 2.6, 05/15/2024
}
\examples{
ukgas_seas <- seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)",
                             x11="", transform.function = "log", forecast.maxlead=20,
                             check.print = c( "pacf", "pacfplot" ))
ukgas_udg <- seasonal::udg(ukgas_seas)
ukgas_acf_warn <- acf_warn(ukgas_udg, acf_lags_warn = c(4,8))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
