% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_warn_why.r
\name{acf_warn_why}
\alias{acf_warn_why}
\title{ACF Test Warning Message}
\usage{
acf_warn_why(udg_list = NULL, acf_lags_warn = c(12, 24), return_both = FALSE)
}
\arguments{
\item{udg_list}{\itemize{
\item list object generated by \code{udg()} function of the \code{seasonal} package.
This is a required entry.
}}

\item{acf_lags_warn}{\itemize{
\item lags of the ACF to test for warnings
Default is \code{c(12, 24)}.
}}

\item{return_both}{Logical scalar indicating whether the calling function will return both the test results and why the test failed or produced a warning. Default is FALSE.}
}
\value{
Character string which tells why the series generates a warning for the ACF test, 'pass' otherwise.
}
\description{
Generates text on why the sample autocorrelation of the residuals from a time series model fails the Ljung-Box or Box-Pierce Q test
}
\details{
Version 3.5, 5/25/2023
}
\examples{
ukgas_seas <- seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)",
                             x11="", transform.function = "log", forecast.maxlead=20,
                             check.print = c( "pacf", "pacfplot" ))
ukgas_udg <- seasonal::udg(ukgas_seas)
ukgas_acf_warn_why <- acf_warn_why(ukgas_udg, acf_lags_warn = c(4, 8), return_both = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
