% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_model_diag.r
\name{all_model_diag}
\alias{all_model_diag}
\title{Model diagnostic summary}
\usage{
all_model_diag(
  seas_obj = NULL,
  add_aicc = FALSE,
  add_norm = FALSE,
  add_auto_out = FALSE,
  add_seasonal = FALSE,
  add_spec = FALSE,
  add_lbq = TRUE,
  set_lbq_lags_fail = c(12, 24),
  set_lbq_p_limit = 0.01,
  set_acf_num_sig = 8,
  set_acf_lags_fail = c(1, 2, 3, 4, 12, 24),
  set_acf_lags_warn = c(12, 24),
  add_pacf = TRUE,
  qs_test_span = FALSE,
  set_qs_p_limit_pass = 0.01,
  set_qs_p_limit_warn = 0.05,
  set_qs_robust_sa = TRUE,
  set_spec_peak_level = 6,
  set_spec_peak_warn = 3,
  return_list = FALSE
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{add_aicc}{logical scalar; add AICC value to the summary.
Default is \code{FALSE}}

\item{add_norm}{logical scalar; add normality statistics to the summary.
Default is \code{FALSE}}

\item{add_auto_out}{logical scalar; add identified automatic outliers to the summary.
Default is \code{FALSE}}

\item{add_seasonal}{logical scalar; add QS test for seasonality to the summary.
Default is \code{FALSE}.}

\item{add_spec}{logical scalar; add test for spectral peaks to the summary.
Default is \code{FALSE}.}

\item{add_lbq}{logical scalar; add test for Ljung-Box Q to the summary.
Default is \code{TRUE}; if set to \code{FALSE}, a more extensive test of ACF
will be done.}

\item{set_lbq_lags_fail}{\itemize{
\item lags of the Ljung-Box Q to test
Default is \code{c(12, 24)}.
}}

\item{set_lbq_p_limit}{Numeric scalar; P-value limit for Ljung-Box Q test
Default is \code{0.01}.}

\item{set_acf_num_sig}{Limit for number of lags with significant ACF values
Default is \code{8}.}

\item{set_acf_lags_fail}{\itemize{
\item lags of the ACF to test
Default is \code{c(1, 2, 3, 4, 12, 24)}.
}}

\item{set_acf_lags_warn}{\itemize{
\item lags of the ACF to test for warnings
Default is \code{c(12, 24)}.
}}

\item{add_pacf}{logical scalar; include PACF in ACF results.
Default is \code{TRUE}.}

\item{qs_test_span}{logical scalar; test span of data in QS seasonal test rather than full series.
Default is \code{FALSE}}

\item{set_qs_p_limit_pass}{Numeric scalar; P-value limit for QS statistic for passing
Default is \code{0.01}.}

\item{set_qs_p_limit_warn}{Numeric scalar; P-value limit for QS statistic for warning
Default is \code{0.05}.}

\item{set_qs_robust_sa}{Logical scalar indicating if original series adjusted for extremes is
included in testing. Default is \code{TRUE}.}

\item{set_spec_peak_level}{Integer scalar - limit to determine if a frequency has a spectral peak.
Default is \code{6}.}

\item{set_spec_peak_warn}{Integer scalar - limit to produce a warning that a frequency may have a
spectral peak. Default is \code{3}.}

\item{return_list}{logical scalar; return a list rather than a vector.
Default is \code{FALSE}}
}
\value{
vector or list of model diagnostics for a given series
}
\description{
Generate a summary of model diagnostics for a single series
}
\details{
Version 5.5, 6/21/2024
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, x11="", slidingspans = "", 
                  transform.function = "log", arima.model = "(0 1 1)(0 1 1)", 
                  regression.aictest = "td", forecast.maxlead=36, 
                  check.print = c( "pacf", "pacfplot" ), check.save = "acf")
air_diag <- 
    all_model_diag(air_seas, add_seasonal = TRUE, add_aicc = TRUE, add_norm = TRUE, 
                   add_auto_out = TRUE, add_pacf = FALSE, qs_test_span = TRUE, 
                   return_list = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
