% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stats.r
\name{check_stats}
\alias{check_stats}
\title{Displays various X-13 diagnostics}
\usage{
check_stats(
  seas_obj = NULL,
  print_summary = TRUE,
  test_full = TRUE,
  test_span = TRUE,
  acf_num_sig = 8,
  acf_lags_fail = c(1, 2, 3, 4, 12, 24),
  acf_lags_warn = c(12, 24),
  model_t_value = 3,
  model_p_value = 0.05,
  otl_auto_limit = 5,
  otl_all_limit = 5,
  d11f_p_level = 0.01,
  qs_p_limit_pass = 0.01,
  qs_p_limit_warn = 0.05,
  qs_p_limit_fail = 0.01,
  qs_robust_sa = TRUE,
  sf_limit = 25,
  change_limit = 40,
  mq_fail_limit = 1.2,
  mq_warn_limit = 0.8,
  return_list = FALSE
)
}
\arguments{
\item{seas_obj}{object generated by \code{seas()} of the seasonal package.}

\item{print_summary}{Logical object; if \code{TRUE}, print the result of summary(\code{seas_obj});
if \code{FALSE}, a model summary will be printed out.
Default is \code{TRUE}.}

\item{test_full}{Logical scalar indicating whether to apply the QS test to the full series span.
Default is \code{TRUE}.}

\item{test_span}{Logical scalar indicating whether to test the QS test to the final 8-year span
used by the spectrum diagnostic.
Default is \code{TRUE}.}

\item{acf_num_sig}{Numeric object; limit for number of lags with significant ACF values.
Default is \code{8}.}

\item{acf_lags_fail}{Numeric vector; lags of the ACF to test.
Default is \code{c(1, 2, 3, 4, 12, 24)}.}

\item{acf_lags_warn}{Numeric vector; lags of the ACF to test for warnings,
Default is \code{c(12, 24)}.}

\item{model_t_value}{t-statistic limit for regressors.
Default is \code{3.0}.}

\item{model_p_value}{p-value limit for regressors.
Default is \code{0.05}.}

\item{otl_auto_limit}{limit for number of automatically identified outliers.
Default is \code{5}.}

\item{otl_all_limit}{limit for number of outlier regressors.
Default is \code{5}.}

\item{d11f_p_level}{p-level used to test the d11 f-test for residual seasonality.
Default is \code{0.01}.}

\item{qs_p_limit_pass}{Numeric scalar; P-value limit for QS statistic for passing.
Default is \code{0.01}.}

\item{qs_p_limit_warn}{Numeric scalar; P-value limit for QS statistic for warning.
Default is \code{0.05}.}

\item{qs_p_limit_fail}{Numeric scalar; P-value limit for QS statistic for failing.
Default is \code{0.01}.}

\item{qs_robust_sa}{Logical scalar indicating if original series adjusted for extremes
is included in testing.
Default is \code{TRUE}.}

\item{sf_limit}{Numeric object; limit for the percentage of seasonal spans flagged.
Default is \code{25}.}

\item{change_limit}{Numeric object; limit for the percentage of month-to-month changes flagged
Default is \code{40}.}

\item{mq_fail_limit}{Numeric scalar; value above which the M or Q statistic fails.
Default is \code{1.2}.}

\item{mq_warn_limit}{Numeric scalar; value above which the M or Q statistic gives a
warning message if it is less than \code{this_fail_Limit},
Default is \code{0.8}.}

\item{return_list}{Logical scalar; indicates if the function will return a summary of
diagnostics. Default is \code{TRUE}.}
}
\value{
Displays assorted seasonal adjusmtent and modeling diagnostics.
}
\description{
Displays various X-13 diagnostics for a single series.
Adapted from a script developed by Osbert Pang (US Census Bureau, osbert.c.pang@census.gov)
}
\details{
Version 3.6, 5/15/2024
}
\examples{
ukgas_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  x11="", transform.function = "log", forecast.maxlead=20,
                  check.print = c( "pacf", "pacfplot" ))
ukgas_check_stats <- 
   check_stats(ukgas_seas, acf_num_sig = 5, acf_lags_fail = c(1, 2, 3, 4, 8), 
               acf_lags_warn = c(4, 8), otl_auto_limit = 4, otl_all_limit = 6, 
               return_list = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
