% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_optimal_seasonal_filter.r
\name{choose_optimal_seasonal_filter}
\alias{choose_optimal_seasonal_filter}
\title{Choose Optimal X-11 seasonal moving average}
\usage{
choose_optimal_seasonal_filter(
  this_seasonal_theta = NULL,
  dp_limits = TRUE,
  use_3x15 = TRUE
)
}
\arguments{
\item{this_seasonal_theta}{numeric scalar; seasonal moving average coefficient from an
airline model. This is a required entry.}

\item{dp_limits}{logical scalar, if \code{TRUE} limits from Deputot and Planas will be used
to choose the moving average, else limits from Bell Chow and Chu will be used.
Default is \code{TRUE}.}

\item{use_3x15}{logical scalar, if \code{TRUE} 3x15 seasonal filter will be returned if chosen,
otherwise function will return a \code{3x9} value.
Default is \code{FALSE}.}
}
\value{
The optimal X-11 seasonal filter, unless the airline model cannot be estimated. Uses the
result from \code{get_seasonal_theta}; if this isn't a null value, return choice of
optimal seasonal moving average
}
\description{
Choose the optimal X-11 seasonal moving average based on the value of the seasonal moving
average coefficient from an airline model.
}
\details{
Version 1.4, 5/23/2023
}
\examples{
shoes_seas <- seasonal::seas(shoes2008, x11="", slidingspans = "", 
                        transform.function = "log", x11 = "",
                        arima.model = "(0 1 1)(0 1 1)", 
                        regression.aictest = c("td", "easter"),
                        forecast.maxlead=36, check.print = c( "pacf", "pacfplot" ))
shoes_seasonal_MA <- shoes_seas$est$coefficients[["MA-Seasonal-12"]]
this_seasonal  <- choose_optimal_seasonal_filter(shoes_seasonal_MA)
this_seasonal2 <- choose_optimal_seasonal_filter(shoes_seasonal_MA, dp_limits = FALSE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
