% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_spectrum_test.r
\name{combined_spectrum_test}
\alias{combined_spectrum_test}
\title{Combined spectrum test from Maravall (2012)}
\usage{
combined_spectrum_test(
  seas_obj = NULL,
  this_ar_spec_cv = NULL,
  this_series = "series.adjoriginal",
  take_log = TRUE,
  take_diff = TRUE
)
}
\arguments{
\item{seas_obj}{\code{seas} object for a single series. This is a required entry.}

\item{this_ar_spec_cv}{List object with two elements - 99 and 95 percent critical values for
the frequencies of the AR(30) spectrum as generated by the \code{gen_ar_spec_cv}}

\item{this_series}{character string; the table used to generate the AR(30) spectrum.
Default is \code{"b1"}.}

\item{take_log}{logical scalar; indicates if the AR spectrum is generated from the log of
the data. Default is \code{TRUE}.}

\item{take_diff}{logical scalar; indicates if the data is differenced before the AR spectrum
is generated. Default is \code{TRUE}.}
}
\value{
\code{TRUE} if spectral evidence of seasonality is detected; \code{FALSE} if not.
}
\description{
Generate a test for seasonality by combining the results from the AR(30) and Tukey nonparametric
spectrums as laid out in Maravall (2012)
}
\details{
Version 2.7, 5/14/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, series.save = "b1",
                    arima.model="(0 1 1)(0 1 1)",
                    forecast.maxlead = 36, slidingspans = "",
                    transform.function = "log")
this_ar30_spec_cv <- gen_ar_spec_cv(1000, 97, 12)
this_spectrum_test <- combined_spectrum_test(air_seas, this_ar30_spec_cv)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
