% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dates.r
\name{compare_dates}
\alias{compare_dates}
\title{Date Match}
\usage{
compare_dates(this_date = NULL, comp_date = NULL)
}
\arguments{
\item{this_date}{Integer array of length 2, a date where the first element is the year and
the second element is the month or quarter. This is a required entry.}

\item{comp_date}{Integer array of length 2, a date to comapare to \code{this_date}.}
}
\value{
Logical scalar; \code{TRUE} if the dates match, \code{FALSE} if they don't.
}
\description{
Compare two dates to see if they match.
}
\details{
Version 3.2, 5/23/2023
}
\examples{
match_start <- compare_dates(start(shoes2007), c(1990,1))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
