% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_string_to_date.r
\name{convert_date_string_to_date}
\alias{convert_date_string_to_date}
\title{Convert date string from UDG output}
\usage{
convert_date_string_to_date(this_date_string)
}
\arguments{
\item{this_date_string}{date string usually extracted from the X-13 UDG output}
}
\value{
integer array of length 2 with the year and month/quarter of from the date string
}
\description{
convert a date string from the X-13 UDG file to a \code{c(year, month)} date
}
\details{
Version 1.1, 4/22/2021
}
\examples{
air_seas <- seasonal::seas(AirPassengers, 
                    arima.model="(0 1 1)(0 1 1)", 
                    forecast.maxlead = 36, slidingspans = "", 
                    transform.function = "log")
this_start_spec_string <- seasonal::udg(air_seas, "startspec")
this_start_spec        <- convert_date_string_to_date(this_start_spec_string)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
