% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d11f_test.r
\name{d11f_test}
\alias{d11f_test}
\title{D11 F-test for residual seasonality}
\usage{
d11f_test(seas_obj = NULL, p_level = 0.01, return_this = "test")
}
\arguments{
\item{seas_obj}{Object generated by \code{seas()} of the seasonal package.
This is a required entry.}

\item{p_level}{p-level used to test the d11 f-test for residual seasonality
Default is \code{0.01}.}

\item{return_this}{character string; what the function returns -
\code{'test'} returns test results, \code{'why'} returns why the test failed or
received a warning, or \code{'both'}. Default is \code{'test'}.}
}
\value{
A text string denoting if series passes or has a warning for residual seasonality.
If d11f statistic not found, return \code{'none'}.
}
\description{
Generates X-11's f-test for residual seasonality in the seasonally adjusted data.
}
\details{
Version 4.3, 5/14/2024
}
\examples{
ukgas_seas <- seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                             x11="", transform.function = "log", forecast.maxlead=20,
                             check.print = c( "pacf", "pacfplot" ))
ukgas_d11f_test <- d11f_test(ukgas_seas, p_level = 0.05, return_this = 'both')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
