% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d11f_test_why.r
\name{d11f_test_why}
\alias{d11f_test_why}
\title{ACF Test Warning Message}
\usage{
d11f_test_why(udg_list = NULL, p_level = 0.01, return_both = FALSE)
}
\arguments{
\item{udg_list}{List object generated by \code{udg()} function of the \code{seasonal} package.
This is a required entry.}

\item{p_level}{p-level used to test the d11 f-test for residual seasonality
Default is \code{0.01}.}

\item{return_both}{Logical scalar indicating whether the calling function will return both
the test results and why the test failed or produced a warning. Default is \code{FALSE}.}
}
\value{
A text string denoting why a series fails or has a warning for residual seasonality.
If d11f statistic not found, return \code{'none'}.
}
\description{
Shows why D11 f-test for residual seasonality fails.
}
\details{
Version 4.2, 5/24/2023
}
\examples{
ukgas_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  x11="", transform.function = "log", forecast.maxlead=20,
                  check.print = c( "pacf", "pacfplot" ))
ukgas_udg <- seasonal::udg(ukgas_seas)
ukgas_d11f_why <- d11f_test_why(ukgas_udg, p_level = 0.05, return_both = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
