% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_dates.r
\name{difference_dates}
\alias{difference_dates}
\title{Date difference}
\usage{
difference_dates(first_date = NULL, second_date = NULL, this_frequency = 12)
}
\arguments{
\item{first_date}{Integer array of length 2, a date where the first element is the year and
the second element is the month or quarter. This is a required entry.}

\item{second_date}{Integer array of length 2, a date to comapare to \code{this_date}.
This is a required entry.}

\item{this_frequency}{Integer scalar, frequency of target time series.
Entries limited to 1 (yearly), 2 (biannual), 3 (triannual), 4 (quarterly),
6 (Bimonthly), and 12. Default is 12 (monthly).}
}
\value{
Integer scalar; difference between first date and second date.
}
\description{
Generate the difference between two dates.
}
\details{
Version 1.4, 12/31/2024
}
\examples{
diff_start <- difference_dates(start(shoes2007), c(1990,1))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
