% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_diag_list.r
\name{fix_diag_list}
\alias{fix_diag_list}
\title{Fix Diagnostic List}
\usage{
fix_diag_list(this_test = NULL, this_names = NULL, return_this = "both")
}
\arguments{
\item{this_test}{list object of a seasonal adjustment or modeling diagnostic
This is a required entry.}

\item{this_names}{character vector; complete set of names to check against
This is a required entry.}

\item{return_this}{character string; what the function returns -
\code{'test'} returns test results, \code{'why'} returns why the test failed or
received a warning, or \code{'both'}. Default is \code{'both'}.}
}
\value{
diagnostic list object with missing names filled in
}
\description{
Fix an incomplete diagnostic list by filling in missing elements with NAs
}
\details{
Version 1.6, 5/25/2023
}
\examples{
xt_composite_seas <- seasonal::seas(xt_data_new,
     transform.function = "log", 
     check.print = c("none", "+acf", "+acfplot", "+normalitytest"), 
     regression.aictest = NULL, 
     outlier.types = "all",
     arima.model = "(0 1 1)(0 1 1)",
     list = list(
         list(x11 = ""),
         list(x11 = ""),
         list(seats.save = c("s11", "s12", "s10")),
         list(x11 = "")
     ),
)
xt_comp_update <- 
     Filter(function(x) inherits(x, "seas"), xt_composite_seas)
xt_test_m7 <- lapply(xt_comp_update, function(x) 
     try(mq_test(x, return_this = 'both')))
test_names <- names(xt_data_new)
num_names <- length(test_names)
if (!is.null(xt_test_m7)) {
    if (length(xt_test_m7) < num_names) {
      xt_test_m7 <- 
         fix_diag_list(xt_test_m7, test_names, return_this = 'both')
    }
}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
