% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ar_spec_cv.r
\name{gen_ar_spec_cv}
\alias{gen_ar_spec_cv}
\title{AR(30) Spectrum Critical Value}
\usage{
gen_ar_spec_cv(n_sim = 1e+05, series_length = 121, freq = 12)
}
\arguments{
\item{n_sim}{integer scalar; number of simulations; default is 100000}

\item{series_length}{integer scalar; length of each series simulated.
Default is 121.}

\item{freq}{integer scalar; frequency of the time series; default is 12 (monthly).}
}
\value{
List of critical values for each seasonal frequency for the 95th and 99th percentile.
}
\description{
Generate critical values for AR(30) spectrum as in Maravall (2012).
}
\details{
Version 1.5, 5/25/2023
}
\examples{
ar30_spec_cv <- gen_ar_spec_cv(1000, 97, 12)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
