% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hybrid_sa.r
\name{gen_hybrid_sa}
\alias{gen_hybrid_sa}
\title{Generate a hybrid seasonal adjustment}
\usage{
gen_hybrid_sa(
  this_mult_sa = NULL,
  this_add_sa = NULL,
  this_start_hybrid = NULL,
  this_end_hybrid = NULL
)
}
\arguments{
\item{this_mult_sa}{time series object of a multiplicative seasonal adjustment
This is a required entry.}

\item{this_add_sa}{time series object of an additive seasonal adjustment
This is a required entry.}

\item{this_start_hybrid}{integer vector of length 2, start of the span where additive
adjustments replace multiplicative adjustment. This is a required entry.}

\item{this_end_hybrid}{integer vector of length 2, end of the span where additive adjustments
replace multiplicative adjustment. This is a required entry.}
}
\value{
time series object with hybrid seasonal adjustment.
}
\description{
Generates a "hybrid" seasonal adjustment by replacing a span of a multiplicative seasonal
adjustment with an additive adjustment.
}
\details{
Version 2.2, 5/25/2023
}
\examples{
air_mult_seas <- seasonal::seas(AirPassengers, transform.function = "log")
air_mult_sa   <- seasonal::final(air_mult_seas)
air_add_seas  <- seasonal::seas(AirPassengers, transform.function = "none")
air_add_sa    <- seasonal::final(air_add_seas)
air_hybrid_sa <- gen_hybrid_sa(air_mult_sa, air_add_sa, c(1956,1), c(1956,12))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
