% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ls_outlier_ts.r
\name{gen_ls_outlier_ts}
\alias{gen_ls_outlier_ts}
\title{Level change regression variable as a \code{ts} object}
\usage{
gen_ls_outlier_ts(
  ls_date,
  this_start,
  this_end,
  this_freq = 12,
  x13type = TRUE,
  return_matrix = TRUE
)
}
\arguments{
\item{ls_date}{Integer vector of length two - dates for LS outlier to be generated.}

\item{this_start}{Numeric vector; start date of LS outlier regressor generated.}

\item{this_end}{Numeric vector; end date of LS outlier regressor generated.}

\item{this_freq}{Numeric scalar; frequency of time series.
Default: \code{12}, for a monthly series}

\item{x13type}{Logical scalar; Indicates if level change outlier is defined as in X-13ARIMA-SEATS.
Default: \code{TRUE}}

\item{return_matrix}{Logical scalar; If true, the object returned is a one column time series
matrix object. Default: \code{TRUE}}
}
\value{
Generate \code{ts} object of a level change outlier regressor
}
\description{
Generates a ts object for a LS (level shift) outlier regressor.
}
\details{
Version 1.8, 5/23/2023
}
\examples{
UKgas_ls_date <- c(1970, 2)
UKgas_ls_1970_2 <-
   gen_ls_outlier_ts(UKgas_ls_date, this_start = c(1960, 1), this_end = c(1990, 4), 
                     this_freq = 4)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
