% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tc_outlier_ts.r
\name{gen_tc_outlier_ts}
\alias{gen_tc_outlier_ts}
\title{Temporary change outlier regression as a \code{ts} object}
\usage{
gen_tc_outlier_ts(
  tc_date = NULL,
  this_start = NULL,
  this_end = NULL,
  this_freq = 12,
  tc_alpha = NULL,
  return_matrix = TRUE
)
}
\arguments{
\item{tc_date}{Integer vector of length two - dates for TC outlier to be generated
This is a required entry.}

\item{this_start}{Numeric vector; start date of TC outlier regressor generated.
This is a required entry.}

\item{this_end}{Numeric vector; end date of TC outlier regressor generated.
This is a required entry.}

\item{this_freq}{Numeric scalar; frequency of time series.
Default: \code{12}, for a monthly series}

\item{tc_alpha}{Numeric scalar; Rate of decay for the TC outlier.
Default: will be computed as in \code{X-13ARIMA-SEATS} for a weekly series}

\item{return_matrix}{Logical scalar; If true, the object returned is a one column
time series matrix object. Default: \code{TRUE}}
}
\value{
\code{ts} object for a temporary change outlier regressor
}
\description{
Generates a \code{ts} object for a TC (temporary change) outlier regressor
}
\details{
Version 1.7, 5/25/2023
}
\examples{
UKgas_tc_date <- c(1970, 2)
UKgas_tc_1970_2 <-
   gen_tc_outlier_ts(UKgas_tc_date, this_start = c(1960, 1), this_end = c(1990, 4), 
                     this_freq = 4, tc_alpha = 0.9, return_matrix = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
