% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_x13_table_list.r
\name{gen_x13_table_list}
\alias{gen_x13_table_list}
\title{X-13 Tables Available}
\usage{
gen_x13_table_list(this_table_type = "all")
}
\arguments{
\item{this_table_type}{vector of character strings listing types of X-13 tables to output.
Default is \code{'all'}, other choices are \code{'diagnostics'}, \code{'matrices'},
\code{'spectrum'}, \code{'timeseries'}.}
}
\value{
A list of arrays with table names and abbreviations from X-13ARIMA-SEATS in
several different elements specified by the user: diagnostics, matrices, spectrum,
timeseries
}
\description{
Generates a list of X-13 tables that can be extracted with the \code{seasonal} package.
}
\details{
Version 2.4, 05/24/2023
}
\examples{
x13_tables_all <- gen_x13_table_list()
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
