% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acf.r
\name{get_acf}
\alias{get_acf}
\title{get ACF values, diagnostics}
\usage{
get_acf(
  seas_obj = NULL,
  lags = NULL,
  as_data_frame = TRUE,
  add_rownames = TRUE
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated by the \code{seasonal} package.
This is a required entry.}

\item{lags}{An array of integer lags to extract from the ACF output.
If not specified, all lags will be returned.}

\item{as_data_frame}{Logical scalar. If \code{TRUE}, the values are returned as a data frame.
If \code{FALSE}, the result is returned as a matrix object. Default is \code{TRUE}.}

\item{add_rownames}{Logical scalar.
If \code{TRUE}, row names are constructed from the lags specified by the user.
If \code{FALSE}, no rownames will be added. Default is \code{TRUE}.}
}
\value{
A data frame (or matrix) with 5 columns:  \code{"Sample_ACF"} (sample ACF Value),
\code{"SE_of_ACF"} (ACF standard error), \code{"Ljung.Box_Q"} (Ljung-Box Q statistic),
\code{"df_of_Q"} (degrees of freedom for Ljung=Box Q), and \code{"P.value"} (p-value for Ljung Box Q).
}
\description{
Extract values from the sample ACF of the regARIMA model residuals, along with Ljung-Box Q values.
}
\details{
Version 2.0, 5/21/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers,
                    arima.model="(0 1 1)(0 1 1)",
                    check.save = "acf",
                    check.maxlag = 36,
                    forecast.maxlead = 36, slidingspans = "",
                    transform.function = "log")
air_acf_df <- get_acf(air_seas, lags = c(1,2,3,12,13,24))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
