% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arima_estimates_matrix.r
\name{get_arima_estimates_matrix}
\alias{get_arima_estimates_matrix}
\title{ARMA Coefficient Summary}
\usage{
get_arima_estimates_matrix(seas_obj = NULL, add_diff = FALSE)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series.
This is a required entry.}

\item{add_diff}{logical scalar; add differencing information, if included in model}
}
\value{
matrix of ARMA coefficients, standard errors, and t-statistics for a given series
}
\description{
Generate a summary of ARMA coefficients for a single series.
}
\details{
Version 3.2, 5/14/2024
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, x11="", slidingspans = "", transform.function = "log",
                        arima.model = "(0 1 1)(0 1 1)", regression.aictest = 'td', 
                        forecast.maxlead=36, check.print = c( "pacf", "pacfplot" ))
air_arima_matrix <- get_arima_estimates_matrix(air_seas, add_diff = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
