% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fcst_tval.r
\name{get_fcst_tval}
\alias{get_fcst_tval}
\title{T-values of within sample forecasts}
\usage{
get_fcst_tval(seas_obj = NULL, terror_lags = NULL)
}
\arguments{
\item{seas_obj}{\code{seas} object for a single series
This is a required entry.}

\item{terror_lags}{Integer scalar for number of forecast lags from the end of series we'll
collect t-statistics. Must be either 1, 2, or 3.
This is a required entry.}
}
\value{
A numeric array of t-values of within sample forecasts, up to length 3.
}
\description{
Returns t-values of within sample forecasts, up to 3
}
\details{
Version 2.10, 5/14/2024
}
\examples{
air_seas_short <- seasonal::seas(AirPassengers, series.span = ',1960.9', 
                    arima.model="(0 1 1)(0 1 1)", x11 = "",
                    forecast.maxlead = 36, slidingspans = "", 
                    transform.function = "log")
fcst_tstat <- get_fcst_tval(air_seas_short, 3)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
