% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ftest_from_udg.r
\name{get_ftest_from_udg}
\alias{get_ftest_from_udg}
\title{Get F-test info from external UDG file}
\usage{
get_ftest_from_udg(this_base = NULL, this_path = NULL, this_reg = "Seasonal")
}
\arguments{
\item{this_base}{Character scalar; Filename of output file of the X-13 run.
This is a required entry.}

\item{this_path}{Character scalar; Path of \code{udg} file.
This is an optional entry.}

\item{this_reg}{Character scalar; Type of regressor for model-based F-test.
Default is \code{Seasonal}.}
}
\value{
Numeric vector of the degrees of freedom, F-test, and p-value for this
model based F-test
}
\description{
Parse \code{udg} file and get information for model-based F-test for a regressor.
}
\details{
Version 1.1, 1/10/2024
}
\examples{
\dontrun{this_ftest<-get_ftest_from_udg("AE1011330000_auto", "X:/code/census_build_60/basic/",
                                        "Trigonometric Seasonal")}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
