% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_ftest.r
\name{get_model_ftest}
\alias{get_model_ftest}
\title{Get model based F-test}
\usage{
get_model_ftest(seas_obj = NULL, this_ftest = "seasonal", return_this = "all")
}
\arguments{
\item{seas_obj}{A seas object for a single series generated from the \code{seasonal} package
This is a required entry.}

\item{this_ftest}{Character string; type of model based f-test to return.
Default is \code{"seasonal"}; only other acceptable value is \code{"td"}.}

\item{return_this}{Character string, Code that controls what values are returned.
Acceptable values are \code{"all"}, \code{"dof"} (degrees of freedom), \code{"ftest"},
(F-test value), or \code{"pval"} (F-test p-value). Default is \code{"all"}.}
}
\value{
Numeric vector with seasonal or trading day F-statistic, degrees of freedom, p-value.
If not found, return \code{NULL}.
}
\description{
Extract values associated with the model based F-test specified by the \code{this_ftest} argument.
}
\details{
Version 3.8, 5/15/2024
}
\examples{
air_seas_td_seasonal <- 
    seasonal::seas(AirPassengers, arima.model = "(0 1 1)", 
                   regression.variables = c("seasonal", "td"), x11="")
this_seas_ftest <- get_model_ftest(air_seas_td_seasonal, this_ftest = "seasonal", 
                                   return_this = "ftest")
this_td_ftest   <- get_model_ftest(air_seas_td_seasonal, this_ftest = "td", 
                                   return_this = "ftest")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
