% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mq_key.r
\name{get_mq_key}
\alias{get_mq_key}
\title{Make a UDG key for X-11-ARIMA M and Q statistics}
\usage{
get_mq_key(this_label = NULL)
}
\arguments{
\item{this_label}{character string; name of an \code{X-11-ARIMA} M and Q statistics
This is a required entry.}
}
\value{
character string with the corresponding UDG label for \code{this_label}.
If incorrect label is specified, returns \code{NULL}.
}
\description{
Generates the UDG key for \code{X-11-ARIMA} M and Q statistics based on a label
}
\details{
Version 1.3, 05/24/2023
}
\examples{
m7_key  <- get_mq_key('M7')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
