% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mq_label.r
\name{get_mq_label}
\alias{get_mq_label}
\title{Make a label for X-11-ARIMA M and Q statistics}
\usage{
get_mq_label(this_key = "f3.q")
}
\arguments{
\item{this_key}{character string; name of an X-11-ARIMA M and Q statistics used in the
UDG X-13 output. Default is \code{"f3.q"}.}
}
\value{
character string with the corresponding label for \code{this_key}.
If incorrect label is specified, returns \code{NULL}.
}
\description{
Generates a label for X-11-ARIMA M and Q statistics
}
\details{
Version 1.2, 05/24/2023
}
\examples{
m7_label  <- get_mq_label('f3.m07')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
