% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nonseasonal_theta.r
\name{get_nonseasonal_theta}
\alias{get_nonseasonal_theta}
\title{Nonseasonal Moving Average from Airline Model}
\usage{
get_nonseasonal_theta(
  seas_obj = NULL,
  this_index = 1,
  return_string = TRUE,
  significant_digits = 3
)
}
\arguments{
\item{seas_obj}{A seas object for a single series generated from the \code{seasonal} package.
This is a required entry.}

\item{this_index}{An integer scalar, an index of the vector values to be passed.
Acceptable values are \code{1} (nonseasonal MA coefficient value),
\code{2} (nonseasonal MA coefficent standard error), or
\code{3} (t-value of the nonseasonal MA coefficient).
Default is \code{1}.}

\item{return_string}{A Logical scalar; indicates whether value returned is a string or numeric.
Default is \code{TRUE}.}

\item{significant_digits}{an integer scalar; significant digits to be saved when a string
is returned. Default is \code{3}.}
}
\value{
Character string containing a value related to the seasonal MA coefficient from the
regARIMA model fit in the \code{seas} object \code{seas_obj}.
If \code{return_string} is \code{FALSE}, this is a numeric.
The standard error or t-value of the seasonal MA coefficient can be returned depending
on the value of \code{this_index}.
}
\description{
Get the value of a nonseasonal moving average coefficient estimated from an airline model.
}
\details{
Version 1.10, May 14, 2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, transform.function = "log", 
                           arima.model = "(0 1 1)(0 1 1)", x11="")
this_nonseasonal_theta <- get_nonseasonal_theta(air_seas, return_string = FALSE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
