% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_norm_stat.r
\name{get_norm_stat}
\alias{get_norm_stat}
\title{Extract normality statistics from X-13}
\usage{
get_norm_stat(seas_obj = NULL, this_norm = NULL)
}
\arguments{
\item{seas_obj}{\code{seas} object generated by the \code{seasonal} package for a single series.
This is a required entry.}

\item{this_norm}{character string; type of normality statistic being extracted.
Permissable values are \code{'a'}, \code{'kurtosis'}, \code{'skewness'}.
This is a required entry.}
}
\value{
Double precision number for normality statistic described in this_key.
If incorrect \code{this_key} used, function returns a \code{NULL} value.
If normality statistic not generated in this run, function returns a \code{NULL} value.
}
\description{
Extract normality statistics from the \code{seas} object of a single series
}
\details{
Version 3.6, 05/14/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, slidingspans = "", transform.function = "log", 
                 x11 = "", forecast.maxlead=36, check.print = c( "pacf", "pacfplot" ))
air_norm_stat <- get_norm_stat(air_seas, this_norm = 'kurtosis')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
