% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outlier_list.R
\name{get_outlier_list}
\alias{get_outlier_list}
\title{List of outliers generated from a list of seasonal objects}
\usage{
get_outlier_list(seas_obj_list = NULL)
}
\arguments{
\item{seas_obj_list}{list of \code{seas} objects generated from a call of \code{seas}
on a single time series. A required argument.}
}
\value{
vector of model diagnostics for a given series
}
\description{
Generate a summary of model outliers from a list of \code{seas} objects series.
}
\details{
Version 2.3, 5/7/2024
}
\examples{
unemp_seas_list <- 
   seasonal::seas(unemployment_list, slidingspans = "", 
                  transform.function = "log", 
                  outlier.types = "all",
                  arima.model = "(0 1 1)(0 1 1)",
                  forecast.maxlead=36, x11 = "",
                  check.print = c( "pacf", "pacfplot" ))
unemp_seas_update <- 
     Filter(function(x) inherits(x, "seas"), unemp_seas_list)
unemp_diag <- 
    get_outlier_list(unemp_seas_update)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
