% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regarima_estimates_matrix.r
\name{get_regarima_estimates_matrix}
\alias{get_regarima_estimates_matrix}
\title{Generate summary of regARIMA model coefficients}
\usage{
get_regarima_estimates_matrix(
  seas_obj = NULL,
  add_diff = FALSE,
  this_xreg_names = NULL
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{add_diff}{logical scalar; add differencing information, if included in model}

\item{this_xreg_names}{Character array; name of user defined regressors.
Default is \code{NULL}, no user defined regressors.
Number of names in this vector should match number of user-defined regressors;
if not, a warning message will be produced.}
}
\value{
A matrix of regARIMA model coefficients, standard errors, and t-statistics
for a given series
}
\description{
Generate a summary of coefficients from a regARIMA model for a single series
}
\details{
Version 2.5, 5/14/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, x11="", slidingspans = "", 
                           transform.function = "log", arima.model = "(0 1 1)(0 1 1)", 
                           regression.aictest = 'td', forecast.maxlead=36, 
                           check.print = c( "pacf", "pacfplot" ))
air_regarima_matrix <- get_regarima_estimates_matrix(air_seas)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
