% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regression_estimates_matrix.r
\name{get_regression_estimates_matrix}
\alias{get_regression_estimates_matrix}
\title{Generate regression coefficient summary}
\usage{
get_regression_estimates_matrix(seas_obj = NULL, this_xreg_names = NULL)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_xreg_names}{Character array; name of user defined regressors.
Default is \code{NULL}, no user defined regressors.}
}
\value{
matrix of regression coefficients, standard errors, and t-statistics for a given series
}
\description{
Generate a summary of regression coefficients for a single series
}
\details{
Version 3.4, 5/14/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, x11="", slidingspans = "", 
                           transform.function = "log", arima.model = "(0 1 1)(0 1 1)", 
                           regression.aictest = 'td', forecast.maxlead=36, 
                           check.print = c( "pacf", "pacfplot" ))
air_reg_matrix <- get_regression_estimates_matrix(air_seas)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
