% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seasonal_ftest_prob.r
\name{get_seasonal_ftest_prob}
\alias{get_seasonal_ftest_prob}
\title{Probability of model based F-test}
\usage{
get_seasonal_ftest_prob(
  seas_obj = NULL,
  this_series = "b1",
  use_seasonal = TRUE
)
}
\arguments{
\item{seas_obj}{\code{seas} object for a single series.
This is a required entry.}

\item{this_series}{character string; the table used to generate the model based F-test.
Default is \code{"b1"}.}

\item{use_seasonal}{logical scalar; if TRUE, the \code{seasonal} regressor is used;
otherwise, use the sine-cosine trignonmetric regressors generated by \code{sincos}.}
}
\value{
test probability generated for the model based seasonal F-test used in the
seasonal testing procedure in Maravall(2012)
}
\description{
Get probability for model based F-test, changing the model to remove seasonal differences
and adding seasonal regressors if necessary. This function is used in the overall seasonal
test from Maravall (2012)
}
\details{
Version 4.4, 6/21/2024
}
\examples{
air_seas_seasonal <- 
  seasonal::seas(AirPassengers, arima.model="(0 1 1)",
                 regression.variables = "seasonal",
                 forecast.maxlead = 36, slidingspans = "", 
                 transform.function = "log", series.save = "a1")
air_ftest_prob <- sautilities::get_seasonal_ftest_prob(air_seas_seasonal)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
