% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timer.r
\name{get_timer}
\alias{get_timer}
\title{Generate elapsed time of run in miliseconds}
\usage{
get_timer(this_base = NULL, this_path = NULL)
}
\arguments{
\item{this_base}{Character scalar; Filename of output file of the X-13 run.
This is a required entry.}

\item{this_path}{Character scalar; Path of profiler output.
This is an optional entry.}
}
\value{
Numeric object of the elapsed time of the X-13 run.
}
\description{
Read profiler information from saved files and generate the elapsed time of an X-13 run in miliseconds.
}
\details{
Version 1.1, 11/29/2023
}
\examples{
\dontrun{AE1011330000_time <- get_timer("AE1011330000_auto", "X:/code/census_build_60/basic/")} 
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
