% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transform.r
\name{get_transform}
\alias{get_transform}
\title{Get transformation}
\usage{
get_transform(seas_obj = NULL)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}
}
\value{
Character string with transformation used to model time series in \code{seas} run
}
\description{
Get transformation from the \code{seas} object of a single time series
}
\details{
Version 1.10, 5/15/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, arima.model = "(0 1 1)(0 1 1)", x11="")
air_trans <- get_transform(air_seas)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
