% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value_from_udg.r
\name{get_value_from_udg}
\alias{get_value_from_udg}
\title{Get value from external UDG file}
\usage{
get_value_from_udg(
  this_base = NULL,
  this_path = NULL,
  this_key = NULL,
  return_numeric = TRUE
)
}
\arguments{
\item{this_base}{Character scalar; Filename of output file of the X-13 run.
This is a required entry.}

\item{this_path}{Character scalar; Path of profiler output.
This is an optional entry.}

\item{this_key}{Character scalar; Key from udg file to search for.
This is a required entry.}

\item{return_numeric}{Logical scalar; Value returned is converted to numeric.
Default is TRUE, returns value as a number.}
}
\value{
Numeric value associated with \code{this_key}.
}
\description{
parse external \code{udg} file and get value for a user supplied key
}
\details{
Version 1.4, 12/12/2023
}
\examples{
\dontrun{this_ftest<-get_ftest_from_udg("AE1011330000_auto", 
                                        "X:/code/census_build_60/basic/",
                                        "aicc")}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
