% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_x11_seasonal_ma.r
\name{get_x11_seasonal_ma}
\alias{get_x11_seasonal_ma}
\title{Get X-ll seasonal moving average}
\usage{
get_x11_seasonal_ma(seas_obj = NULL)
}
\arguments{
\item{seas_obj}{\code{seas} object for a single series,
This is a required entry.}
}
\value{
Returns the seasonal moving average used in an X-ll seasonal adjustment from a seas object.
Returns NULL if X-11 seasonal adjustment is not done.
}
\description{
Returns the seasonal moving average used in an X-ll seasonal adjustment from a seas object.
}
\details{
Version 1.1 3/12/2025
}
\examples{
air_seas_short <- seasonal::seas(AirPassengers, series.span = ',1960.9', 
                    arima.model="(0 1 1)(0 1 1)", 
                    forecast.maxlead = 36, slidingspans = "", 
                    transform.function = "log", x11.save = "d11")
air_sf_short  <- get_x11_seasonal_ma(air_seas_short)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
