% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_est.r
\name{import_est}
\alias{import_est}
\title{Import regCMPNT estimates file}
\usage{
import_est(file_name = NULL, column_name = NULL, return_matrix = TRUE)
}
\arguments{
\item{file_name}{Character string; file name for regCMPNT estimate file.
This is a required entry}

\item{column_name}{Array of character strings; names for the columns of the
estimates matrix. Array must be of length 5.
Default is \code{c("Unscaled_Stochastic", "Scale_Factors", "Scaled_Stochastic", 
       "Regression_Effects", "Combined_Estimate")}.}

\item{return_matrix}{Logical scalar; determines if a matrix or data frame object is returned.
Default is \code{TRUE}.}
}
\value{
A \code{ts} matrix object or a data frame of \code{ts} objects which contains
the contents of the estimates for a given component from a regCMPNT run.
The file name for the for the component file has an \code{.est} file extension.
}
\description{
Reads in an estimated component from a file saved by regCMPNT
}
\details{
Version 1.2, 12/12/2024
}
\examples{
\dontrun{
n3000019_trend_df         <- import_est("n300019_rev2_comp01.est", return_matrix = FALSE)
n3000019_seasonal_df      <- import_est("n300019_rev2_comp02.est", return_matrix = FALSE)
n3000019_irregular_df     <- import_est("n300019_rev2_comp03.est", return_matrix = FALSE)
n3000019_samplingerror_df <- import_est("n300019_rev2_comp04.est", return_matrix = FALSE)
}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
