% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_var.r
\name{import_var}
\alias{import_var}
\title{Import regCMPNT Variance file}
\usage{
import_var(file_name = NULL, column_name = NULL, return_matrix = TRUE)
}
\arguments{
\item{file_name}{Character string; file name for regCMPNT variance file.
This is a required entry}

\item{column_name}{Array of character strings; names for the columns of the
estimates matrix. Array must be of length 4.
Default is \code{c("Unscaled_Stochastic", "Scaled_Stochastic", 
       "Regression_Estimation", "Combined")}.}

\item{return_matrix}{Logical scalar; determines if a matrix object is returned.
Default is \code{TRUE}, which forces the function to return a data frame object.}
}
\value{
A \code{ts} matrix object or a data frame of \code{ts} objects which contains
the contents of the variances for a given component from a regCMPNT run.
The file name for the for the component file has an \code{.var} file extension.
}
\description{
Reads in variances for a component from a file saved by regCMPNT
}
\details{
Version 1.3, 12/12/2024
}
\examples{
\dontrun{
n3000019_trend_var_df         <- import_var("n300019_rev2_comp01.var", return_matrix = FALSE)
n3000019_seasonal_var_df      <- import_var("n300019_rev2_comp02.var", return_matrix = FALSE)
n3000019_irregular_var_df     <- import_var("n300019_rev2_comp03.var", return_matrix = FALSE)
n3000019_samplingerror_var_df <- import_var("n300019_rev2_comp04.var", return_matrix = FALSE)
}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
