% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_saved_x13_file.r
\name{input_saved_x13_file}
\alias{input_saved_x13_file}
\title{Import File Saved by X-13ARIMA-SEATS}
\usage{
input_saved_x13_file(filename = NULL, pos = 2, ncol = 2)
}
\arguments{
\item{filename}{Character string, filename of a file saved by the \code{X-13ARIMA-SEATS}
program. This is a required entry.}

\item{pos}{Integer scalar, column of data to be extracted from \code{filename}.
Default is 2.}

\item{ncol}{Integer scalar, number of columns of data that exist within \code{filename}.
Default is 2.}
}
\value{
A time series array
}
\description{
Import data from a file saved by the X-13ARIMA-SEATS program
}
\details{
Version 1.2, 5/25/2023
}
\examples{
\dontrun{
airline.sa <- input_saved_x13_file("airline.d11")
}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
