% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbq_test.r
\name{lbq_test}
\alias{lbq_test}
\title{Ljung-Box Q ACF test}
\usage{
lbq_test(
  seas_obj = NULL,
  lbq_lags_fail = c(12, 24),
  p_limit = 0.01,
  return_this = "test"
)
}
\arguments{
\item{seas_obj}{Object generated by \code{seas()} of the \code{seasonal} package.
This is a required entry.}

\item{lbq_lags_fail}{\itemize{
\item lags of the ACF to test
Default is \code{c(12, 24)}.
}}

\item{p_limit}{\itemize{
\item numeric limit for the p-value of the Ljung-Box Q
Default is \code{0.01}.
}}

\item{return_this}{character string; what the function returns -
\code{'test'} returns test results, \code{'why'} returns why the test failed or
received a warning, or \code{'both'}. Default is \code{'test'}.}
}
\value{
A text string denoting if series passes, fails, or has a warning for residual
autocorrelation. If Ljung-Box not found, the seasonal object will be updated with
check spec arguments - if this is unsuccessful, return \code{'none'}.
If regARIMA model not estimated, return \code{'none'}
}
\description{
Tests whether the residuals from a time series model has acceptable autocorrelation as indicated
}
\details{
Version 1.7, 5/3/2024
}
\examples{
ukgas_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  x11="", transform.function = "log", forecast.maxlead=20,
                  check.save = 'acf', check.maxlag = 12, check.qlimit = 0.01)
ukgas_lbq_test <- lbq_test(ukgas_seas, lbq_lags_fail = c(4, 8), return_this = 'both')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
