% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_list_number.r
\name{match_list_number}
\alias{match_list_number}
\title{Number of list element matches}
\usage{
match_list_number(this_list, this_string = "fail")
}
\arguments{
\item{this_list}{List of character strings.}

\item{this_string}{Character string to match against elements of the list, ie,
\code{this_string = 'pass'}. Default is \code{'fail'}.}
}
\value{
The number of list items that match \code{this_string}.
}
\description{
Returns number of elements in list that matches \code{this_string}.
}
\details{
Version 3.1, 5/21/2023
}
\examples{
unemp_seas <- 
     seasonal::seas(unemployment_list, x11 = "", 
                    slidingspans = "", 
                    transform.function = "log",
                    arima.model = "(0 1 1)(0 1 1)",
                    forecast.maxlead = 60, 
                    check.print = c( "pacf", "pacfplot" ))
test_seas_update <- 
     Filter(function(x) inherits(x, "seas"), unemp_seas)
test_acf_test <- 
     lapply(test_seas_update, function(x) 
           try(acf_test(x, return_this = 'test')))
test_acf_number_fail <- 
     match_list_number(test_acf_test, 'fail')
test_acf_number_warn <- 
     match_list_number(test_acf_test, 'warn')
test_acf_number_pass <- 
     match_list_number(test_acf_test, 'pass')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
