% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_of_list.r
\name{member_of_list}
\alias{member_of_list}
\title{Member of list}
\usage{
member_of_list(this_list = NULL, this_name = NULL)
}
\arguments{
\item{this_list}{A list of objects. This is a required entry.}

\item{this_name}{character string; element name of \code{this_list}.
This is a required entry.}
}
\value{
returns \code{TRUE} if \code{this_name} is an element of \code{this_list},
\code{FALSE} otherwise
}
\description{
Determines if a name is a member of a list.
}
\details{
Version 1.4, 5/22/2023
}
\examples{
emp_seas_list <- seasonal::seas(employment_list, 
                      slidingspans = "", x11 = "",
                      transform.function = "log",
                      arima.model = "(0 1 1)(0 1 1)",
                      forecast.maxlead=36,
                      check.print = c( "pacf", "pacfplot" ))
emp_seas_update <- 
     Filter(function(x) inherits(x, "seas"), emp_seas_list)
if (member_of_list(emp_seas_update, 'n2000014')) {
    \dontrun{save_spec_file(emp_seas_update$n2000014, 'n2000014',
                  this_directory      = "X:\\\\seasonalAdj\\\\testing\\\\sautilities",
                  this_data_directory = "X:\\\\seasonalAdj\\\\cps_dec_2022\\\\dat",
                  data_file_name      = "n2000014.dat")}
}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
