% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_test.r
\name{model_test}
\alias{model_test}
\title{Tests Time Series Model.}
\usage{
model_test(
  seas_obj = NULL,
  t_value = 3,
  p_value = 0.05,
  otl_auto_limit = 5,
  otl_all_limit = 5,
  return_this = "test"
)
}
\arguments{
\item{seas_obj}{Object generated by \code{seas()} of the \code{seasonal} package.
This is a required entry.}

\item{t_value}{t-statistic limit for regressors. Default is \code{3}.}

\item{p_value}{p-value limit for regressors. Default is \code{0.01}.}

\item{otl_auto_limit}{Numeric object; limit for number of automatically identified outliers.
Default is \code{4}.}

\item{otl_all_limit}{Numeric object; limit for number of outlier regressors.
Default is \code{6}.}

\item{return_this}{character string; what the function returns - \code{'test'}
or \code{'both'}. Default is \code{'both'}.}
}
\value{
A text string denoting if the series passed or failed the tests of ARIMA diagnostics.
}
\description{
Tests whether the time series model has acceptable diagnostics.
}
\details{
Version 4.4, 5/14/2024
}
\examples{
ukgas_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  x11="", transform.function = "log", forecast.maxlead=20,
                  check.print = c( "pacf", "pacfplot" ))
ukgas_model <- model_test(ukgas_seas, t_value=3.0, p_value=0.01, otl_auto_limit=4, 
                          otl_all_limit=6)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
