% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_test_why.r
\name{model_test_why}
\alias{model_test_why}
\title{Model Test Warning Message}
\usage{
model_test_why(
  udg_list = NULL,
  t_value = 3,
  p_value = 0.05,
  otl_auto_limit = 5,
  otl_all_limit = 5,
  return_both = FALSE
)
}
\arguments{
\item{udg_list}{List object generated by \code{udg()} function of the \code{seasonal} package.
This is a required entry.}

\item{t_value}{Numeric scalar; t-statistic limit for regressors. Default is \code{3.0}}

\item{p_value}{Numeric scalar; p-value limit for regressors. Default is \code{0.01}}

\item{otl_auto_limit}{Integer scalar; limit for number of automatically identified outliers.
Default is \code{4}}

\item{otl_all_limit}{Integer scalar; limit for number of automatically identified outliers.
Default is \code{4}}

\item{return_both}{Logical scalar indicating whether the calling function will return
both the test results and why the test failed or produced a warning.
Default is \code{TRUE}.}
}
\value{
A text string denoting why the series passed or failed a series of tests of
ARIMA diagnostics.
ukgas_seas <- seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)",
x11="", transform.function = "log", forecast.maxlead=20,
check.print = c( "pacf", "pacfplot" ))
ukgas_udg <- seasonal::udg(ukgas_seas)
ukgas_model_why <-
model_test_why(ukgas_udg, t_value=3.0, p_value=0.01, otl_auto_limit=4,
otl_all_limit=6, return_both = TRUE)
}
\description{
Generates text on why a time series model is inadequate.
}
\details{
Version 4.2, 5/23/2023
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
