% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_test.r
\name{mq_test}
\alias{mq_test}
\title{Test X-11-ARIMA M and Q statistics}
\usage{
mq_test(
  seas_obj = NULL,
  this_label = "m7",
  this_fail_limit = 1.2,
  this_warn_limit = 0.8,
  return_this = "test"
)
}
\arguments{
\item{seas_obj}{Object generated by \code{seas()} of the seasonal package.
This is a required entry.}

\item{this_label}{Character string; label for an M or Q statistic, such as \code{'M7'}, \code{'Q'},
or \code{'Q2'}. Default is \code{'m7'}.}

\item{this_fail_limit}{Numeric scalar; value above which the M or Q statistic fails.
Default is \code{1.2}.}

\item{this_warn_limit}{Numeric scalar; value above which the M or Q statistic gives a
warning message if it is less than \code{this_fail_Limit}; default is \code{0.8}}

\item{return_this}{character string; what the function returns -
\code{'test'} returns test results, \code{'why'} returns why the test failed or
received a warning, or \code{'both'}. Default is \code{'test'}.}
}
\value{
A text string denoting if series passes or has a warning for residual seasonality.
If \code{D11f} statistic not found, return \code{'none'}.
}
\description{
Generates a test for X-11-ARIMA M and Q statistics.
}
\details{
Version 3.3, 5/14/2024
}
\examples{
ukgas_seas <- seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                             x11="", transform.function = "log", forecast.maxlead=20,
                             check.print = c( "pacf", "pacfplot" ))
ukgas_q <- mq_test(ukgas_seas, this_label = 'q', return_this = 'both')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
