% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_test.r
\name{norm_test}
\alias{norm_test}
\title{Normality Tests for Time Series Models.}
\usage{
norm_test(seas_obj = NULL, this_norm = NULL, return_this = "test")
}
\arguments{
\item{seas_obj}{Object generated by \code{seas()} of the \code{seasonal} package.
This is a required entry.}

\item{this_norm}{Type of normality statistic being extracted;
permissable values are \code{'a'}, \code{'kurtosis'}, \code{'skewness'}.
This is a required entry.}

\item{return_this}{Character string; what the function returns -
\code{'test'} returns test results, \code{'why'} returns why the test failed or
received a warning, or \code{'both'}. Default is \code{'test'}.}
}
\value{
A text string denoting whether the series passed or failed the specific normality test.
If improper value is specified for this_norm, return \code{NULL}.
If no statistic is found, return \code{NA}.
}
\description{
Tests different normality statistics available in X-13ARIMA-SEATS.
}
\details{
Version 4.4, 5/14/2024
}
\examples{
ukgas_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  x11 = "", transform.function = "log", forecast.maxlead = 20,
                  check.print = c( "pacf", "pacfplot" ))
ukgas_norm_a <- norm_test(ukgas_seas, 'a')
ukgas_norm_kurtosis <- norm_test(ukgas_seas, 'kurtosis')
ukgas_norm_skewness <- norm_test(ukgas_seas, 'skewness')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
