% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_test_why.r
\name{norm_test_why}
\alias{norm_test_why}
\title{Normality Test Warning Message}
\usage{
norm_test_why(udg_list = NULL, this_norm = NULL, return_both = FALSE)
}
\arguments{
\item{udg_list}{List object generated by \code{udg()} function of the \code{seasonal} package.
This is a required entry.}

\item{this_norm}{Character scalar inidicating type of normality statistic being extracted;
permissable values are \code{'a'}, \code{'kurtosis'}, \code{'skewness'}.
This is a required entry.}

\item{return_both}{Logical scalar indicating whether the calling function will return
both the test results and why the test failed or produced a warning.
Default is \code{FALSE}.}
}
\value{
A text string showing why a series failed the specific normality test
}
\description{
Generates message for why different normality statistics available in \code{X-13ARIMA-SEATS} fail.
}
\details{
Version 3.2, October 2, 2020
}
\examples{
ukgas_seas <- seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                             x11="", transform.function = "log", forecast.maxlead=20,
                             check.print = c( "pacf", "pacfplot" ))
ukgas_udg <- seasonal::udg(ukgas_seas)
ukgas_norm_a_why <- norm_test_why(ukgas_udg, 'a', return_both = TRUE)
ukgas_norm_kurtosis_why <- norm_test_why(ukgas_udg, 'kurtosis', return_both = TRUE)
ukgas_norm_skewness_why <- norm_test_why(ukgas_udg, 'skewness', return_both = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
