% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP_test.r
\name{NP_test}
\alias{NP_test}
\title{Non-Parametric test from Maravall (2012)}
\usage{
NP_test(x = NULL)
}
\arguments{
\item{x}{\code{ts} time series object. This is a required entry.}
}
\value{
List object with three elements: \code{ken} (test statistic), \code{df}
(degrees of freedom), \code{cv} (test probability)
}
\description{
Non-Parametric test for seasonality based on Kendall and Ord (1990), and
originally due to Friedman from a paper by Maravall.
This code is adapted from \code{kendalls} subroutine in \code{ansub11.f}
from the \code{X-13ARIMA-SEATS} source code
}
\details{
Version 1.7, 5/23/2023
}
\examples{
NP_test_air <- NP_test(AirPassengers)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
