% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_user_reg.r
\name{number_of_user_reg}
\alias{number_of_user_reg}
\title{Number of user defined regressors}
\usage{
number_of_user_reg(seas_obj = NULL)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas}
on a single time series. A required argument.}
}
\value{
number of user defined regressors in the model contained in seas_obj
}
\description{
Extreact the number of user-defined regressors in a \code{seas} objects series.
}
\details{
Version 1.0, 5/1/2025
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, x11="", slidingspans = "", 
                  transform.function = "log", arima.model = "(0 1 1)(0 1 1)", 
                  regression.aictest = "td", forecast.maxlead=36, 
                  check.print = c( "pacf", "pacfplot" ), check.save = "acf")
air_num_user_reg <- number_of_user_reg(air_seas)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
