% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_seasonal_filter.r
\name{optimal_seasonal_filter}
\alias{optimal_seasonal_filter}
\title{Optimal X-11 seasonal moving average selection}
\usage{
optimal_seasonal_filter(
  this_series = NULL,
  aictest = NULL,
  model = "(0 1 1)(0 1 1)",
  variables = NULL,
  outlier = TRUE,
  trans = NULL,
  missing_code = NULL,
  this_xreg = NULL,
  dp_limits = TRUE,
  use_msr = FALSE,
  use_3x15 = TRUE
)
}
\arguments{
\item{this_series}{A time series object. This is a required entry.}

\item{aictest}{a character string with the entries for the \code{regression.aictest} argument
to the \code{seas} function from the \code{seasonal} package.
Default is \code{NULL}, AIC testing not done.}

\item{model}{a character string with the entry for the \code{arima.model} argument to the
\code{seas} function from the \code{seasonal} package.
Default is \code{"(0 1 1)(0 1 1)"}. Model should have a \code{(0 1 1)} seasonal term.}

\item{variables}{a character string with the entries for the \code{regression.variables}
argument to the \code{seas} function from the \code{seasonal} package.
Default is \code{NULL}, no regressors added.}

\item{outlier}{logical scalar, if \code{TRUE} outlier identification is done in the call
to the \code{seas} function from the \code{seasonal} package.
Default is \code{TRUE}.}

\item{trans}{characater scalar, a character string with the entry for the
\code{transform.function} argument to the \code{seas} function.
Default is \code{NULL}, and the entry \code{auto} will be used.}

\item{missing_code}{numeric scalar, a number with the entry for the \code{series.missingcode}
argument to the \code{seas} function.
Default is \code{NULL}, no missing value code is used.}

\item{this_xreg}{numeric matrix, a user defined regressor matrix to be used in the model
estimation. Default is \code{NULL}, no user-defined regressors are used.}

\item{dp_limits}{logical scalar, if \code{TRUE} limits from Deputot and Planas will be used
to choose the moving average, else limits from Bell Chow and Chu will be used.
Default is \code{TRUE}.}

\item{use_msr}{logical scalar, if \code{TRUE} result of MSR selection will be used if model cannot
be estimated, otherwise function will return a \code{NULL} value.
Default is \code{FALSE}.}

\item{use_3x15}{logical scalar, if \code{TRUE} 3x15 seasonal filter will be returned if chosen,
otherwise function will return a \code{3x9} value.
Default is \code{TRUE}.}
}
\value{
The optimal X-11 seasonal filter, unless the airline model cannot be estimated.
}
\description{
Determine the optimal X-11 seasonal moving average based on the value of the seasonal moving
average coefficient from an airline model.
}
\details{
Version 4.3, 5/25/2023
}
\examples{
this_seasonal  <- 
   optimal_seasonal_filter(shoes2008, aictest = c("td", "easter"), use_msr = TRUE)
this_seasonal2 <- 
   optimal_seasonal_filter(shoes2008, aictest = c("td", "easter"), dp_limits = FALSE, 
                           use_msr = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
