% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_seasonal_test_1.r
\name{overall_seasonal_test_1}
\alias{overall_seasonal_test_1}
\title{First overall sasonality test from Maravall (2012)}
\usage{
overall_seasonal_test_1(
  seas_obj = NULL,
  this_series = "a1",
  take_log = TRUE,
  this_ori_qs = "qsori"
)
}
\arguments{
\item{seas_obj}{\code{seas} object for a single series
This argument is required.}

\item{this_series}{character string; the table used to generate the AR(30) spectrum.
Default is \code{"a1"}.}

\item{take_log}{logical scalar; indicates if the AR spectrum is generated from the log
of the data. Default is \code{TRUE}.}

\item{this_ori_qs}{character string; code for which original series QS will be used in this
analysis. Default is \code{"qsori"} (original series, full span); other choices are
\code{"qssorievadj"} (original series adjusted for extreme values,short span),
\code{"qsorievadj"} (original series adjusted for extreme values, full span), and
\code{"qssori"} (original series, short span)}
}
\value{
A list with 3 elements: \code{QStest} (test probability for QS),
\code{NPtest} (test probability for NP),
and \code{result} (character string with test result - possible values of either
\code{"evidence of seasonality"} and \code{"no evidence of seasonality"})
}
\description{
Conduct the first overall test for seasonality as laid out in Maravall (2012)
}
\details{
Version 3.4, 5/21/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers,
                    arima.model="(0 1 1)(0 1 1)",
                    forecast.maxlead = 36, slidingspans = "",
                    transform.function = "log", 
                    series.save = "a1")
first_test <- overall_seasonal_test_1(air_seas)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
