% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_seasonal_test_2.r
\name{overall_seasonal_test_2}
\alias{overall_seasonal_test_2}
\title{Second overall sasonality test from Maravall (2012)}
\usage{
overall_seasonal_test_2(
  seas_obj = NULL,
  this_ar_spec_cv = NULL,
  this_series = "b1",
  this_ori_qs = "qssorievadj",
  take_log = TRUE,
  take_diff = TRUE,
  use_seasonal = TRUE
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated by the \code{seasonal} package.
This is a required entry.}

\item{this_ar_spec_cv}{List object with two elements - 99 and 95 percent critical values for
the frequencies of the AR(30) spectrum as generated by the \code{gen_ar_spec_cv}
function.}

\item{this_series}{character string; the table used to generate the AR(30) spectrum.
Default is \code{"b1"}.}

\item{this_ori_qs}{character string; code for which original series QS will be used in this
analysis. Default is \code{"qssorievadj"} (original series adjusted for extreme values,
short span); other choices are \code{"qsori"} (original series, full span),
\code{"qsorievadj"} (original series adjusted for extreme values, full span), and
\code{"qssori"} (original series, short span)}

\item{take_log}{logical scalar; indicates if the AR spectrum is generated from the log of
the data. Default is \code{TRUE}.}

\item{take_diff}{logical scalar; indicates if the data is differenced before the AR spectrum is
generated. Default is \code{TRUE}.}

\item{use_seasonal}{logical scalar; if TRUE, the \code{seasonal} regressor is used;
otherwise, use the sine-cosine trignonmetric regressors generated by \code{sincos}.}
}
\value{
A list with 5 elements: \code{QStest} (test probability for QS),
\code{NPtest} (test probability for NP),
\code{Ftest} (test probability for model based seasonal F-test),
\code{spectrum} (character string with test result - possible values of either
\code{"evidence of seasonal peak"}, \code{"no evidence of seasonal peak"}),
and \code{result} (character string with test result - possible values of either
\code{"strong seasonal"}, \code{"weak seasonal"}, \code{"no seasonal"}.
}
\description{
Conduct the second overall test for seasonality as laid out in Maravall (2012)
}
\details{
Version 4.1, 5/14/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers,
                    arima.model="(0 1 1)(0 1 1)",
                    series.save = "b1",
                    forecast.maxlead = 36, slidingspans = "",
                    transform.function = "log")
ar30_spec_cv <- gen_ar_spec_cv(1000, 97, 12)
second_test <- overall_seasonal_test_2(air_seas, ar30_spec_cv)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
