% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs_rsd_warn_why.r
\name{qs_rsd_warn_why}
\alias{qs_rsd_warn_why}
\title{Residual QS diagnostic warning message.}
\usage{
qs_rsd_warn_why(
  udg_list = NULL,
  test_full = TRUE,
  test_span = TRUE,
  p_limit_warn = 0.05,
  return_both = FALSE
)
}
\arguments{
\item{udg_list}{List object generated by \code{udg()} function of the \code{seasonal} package.
This is a required entry.}

\item{test_full}{Logical scalar indicating whether to test the full series span.
Default is \code{TRUE}.}

\item{test_span}{Logical scalar indicating whether to test the final 8-year span used by
the spectrum diagnostic. Default is \code{TRUE}.}

\item{p_limit_warn}{Numeric scalar; P-value limit for QS statistic for warning.
Default is \code{0.05}.}

\item{return_both}{Logical scalar indicating whether the calling function will return both
the test results and why the test failed or produced a warning. Default is \code{FALSE}.}
}
\value{
A text string denoting why the series generated a warning message for the QS of
regARIMA residuals.
}
\description{
Generates text explaining why the QS diagnostic failed or generated a warning for regARIMA residuals.
}
\details{
Version 3.3, 5/23/2023
}
\examples{
ukgas_seas <- seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                             x11="", transform.function = "log", forecast.maxlead=20,
                             check.print = c( "pacf", "pacfplot" ))
ukgas_udg <- seasonal::udg(ukgas_seas)
ukgas_qs_rsd_warn <- 
   qs_rsd_warn_why(ukgas_udg, test_full = FALSE, p_limit_warn = 0.05, return_both = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
