% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs_seasonal_fail_why.r
\name{qs_seasonal_fail_why}
\alias{qs_seasonal_fail_why}
\title{QS Test for original series}
\usage{
qs_seasonal_fail_why(
  udg_list = NULL,
  test_full = TRUE,
  test_span = TRUE,
  p_limit_warn = 0.05,
  robust_sa = TRUE,
  return_both = FALSE
)
}
\arguments{
\item{udg_list}{List object generated by \code{udg()} function of the \code{seasonal} package.
This is a required entry.}

\item{test_full}{Logical scalar indicating whether to test the full series span
Default is \code{TRUE}.}

\item{test_span}{Logical scalar indicating whether to test the final 8-year span used
by the spectrum diagnostic. Default is \code{TRUE}.}

\item{p_limit_warn}{Numeric scalar; P-value limit for QS statistic for warning
Default is \code{0.05}.}

\item{robust_sa}{Logical scalar indicating if original series adjusted for extremes is included
in testing. Default is \code{FALSE}.}

\item{return_both}{Logical scalar indicating whether the calling function will return both the}
}
\value{
A text string denoting if the series passed or failed the QS diagnostic.
}
\description{
Tests using the QS diagnostic developed by Maravall to determine if the original series is seasonal
}
\details{
Version 4.2, 5/23/2023
}
\examples{
ukgas_seas <- seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                             x11="", transform.function = "log", forecast.maxlead=20,
                             check.print = c( "pacf", "pacfplot" ))
ukgas_udg <- seasonal::udg(ukgas_seas)
ukgas_qs_fail_seasonal <- 
  qs_seasonal_fail_why(ukgas_udg, test_full = FALSE, 
                       p_limit_warn = 0.05, robust_sa = FALSE, return_both = FALSE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
