% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs_seasonal_test.r
\name{qs_seasonal_test}
\alias{qs_seasonal_test}
\title{QS seasonal tests}
\usage{
qs_seasonal_test(
  seas_obj = NULL,
  test_full = TRUE,
  test_span = TRUE,
  p_limit_pass = 0.01,
  p_limit_warn = 0.05,
  robust_sa = TRUE,
  return_this = "test"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated by the \code{seasonal} package.
This is a required entry.}

\item{test_full}{Logical scalar indicating whether to test the full series span
Default is \code{TRUE}.}

\item{test_span}{Logical scalar indicating whether to test the final 8-year span used by the
spectrum diagnostic. Default is \code{TRUE}.}

\item{p_limit_pass}{Numeric scalar; P-value limit for QS statistic for passing
Default is \code{0.01}.}

\item{p_limit_warn}{Numeric scalar; P-value limit for QS statistic for warning
Default is \code{0.05}.}

\item{robust_sa}{Logical scalar indicating if original series adjusted for extremes is
included in testing. Default is \code{TRUE}.}

\item{return_this}{Character string; what the function returns - \code{'test'} returns
test results, \code{'why'} returns why the test failed or received a warning, or
\code{'both'}. Default is \code{'test'}.}
}
\value{
A text string denoting if the series passed or failed tests for seasonality using
the QS diagnostics. Can test the entire series or the last 8 years or both.
}
\description{
Tests using the QS diagnostic developed by Maravall to determine if the original series is seasonal
}
\details{
Version 4.4, 5/14/2024
}
\examples{
ukgas_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  x11="", transform.function = "log", forecast.maxlead=20,
                  check.print = c( "pacf", "pacfplot" ))
ukgas_qs_test_seasonal <- 
   qs_seasonal_test(ukgas_seas, test_full = FALSE, p_limit_pass = 0.01, 
                    p_limit_warn = 0.05, robust_sa=FALSE, return_this = 'both')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
