% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs_series.r
\name{qs_series}
\alias{qs_series}
\title{QS diagnostic test on a number of series}
\usage{
qs_series(
  seas_obj_list = NULL,
  this_stat = "qsori",
  less_than = TRUE,
  p_limit = 0.01
)
}
\arguments{
\item{seas_obj_list}{list object of \code{seas} object generated by the \code{seasonal}
package. This is a required entry.}

\item{this_stat}{Character string that specifies which QS statistic is being tested.
Allowable values are \code{'qsori'}, \code{'qsorievadj'}, \code{'qsrsd'},
\code{'qssadj'}, \code{'qssadjevadj'}, \code{'qsirr'}, \code{'qsirrevadj'},
\code{'qssori'}, \code{'qssorievadj'}, \code{'qssrsd'}, \code{'qsssadj'},
\code{'qsssadjevadj'}, \code{'qssirr'}, \code{'qssirrevadj'}.
Default is \code{"qsori"}.}

\item{less_than}{Logical scalar which indicates if the test is going to be QS < \code{p_limit}
(\code{less_than} = TRUE) or QS > \code{p_limit} (\code{less_than} = FALSE).
Default is \code{TRUE}.}

\item{p_limit}{Numeric scalar; P-value limit for QS statistic. Default is \code{0.01}.}
}
\value{
A vector of list element names that have the given QS statistic either less than or
greater than the given P-value limit. If nothing matches, the function will output
the string \code{'none'}.
}
\description{
Apply QS Tests to a list of seas objevts
}
\details{
Version 3.3, 5/25/2023
}
\examples{
emp_seas_list <- 
    seasonal::seas(employment_list, 
                   x11 = "", slidingspans = "", 
                   arima.model = "(0 1 1)(0 1 1)",
                   transform.function = "log",
                   regression.aictest = NULL, 
                   forecast.maxlead=60, 
                   check.print = c( "pacf", "pacfplot" ))
test_seas_update <- 
     Filter(function(x) inherits(x, "seas"), emp_seas_list)
test_qs_names <- 
     qs_series(test_seas_update, 'qssori', less_than = FALSE, p_limit=0.025)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
