% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_terror.r
\name{r_terror}
\alias{r_terror}
\title{TERROR for R}
\usage{
r_terror(
  this_series = NULL,
  max_lead = 36,
  log_transform = TRUE,
  aictest = NULL,
  terror_lags = 1
)
}
\arguments{
\item{this_series}{Time series array. This is a required entry.}

\item{max_lead}{Number of forecasts generated by the seas run. Default is 36.}

\item{log_transform}{logical scalar, if TRUE \code{transform.function} will be set to \code{log}
Default is \code{TRUE}.}

\item{aictest}{a character string with the entries for the \code{regression.aictest} argument
to the \code{seas} function from the \code{seasonal} package.
Default is \code{NULL}.}

\item{terror_lags}{Integer scalar for number of forecast lags from the end of series
we'll collect t-statistics. Must be either 1, 2, or 3.}
}
\value{
t-statistics generated by out of sample forecast error for the last 1 to 3
observation of each series in the list.
}
\description{
A function that duplicates the functionality of the TERROR software (Caporello and Maravall 2004)
that performs quality control on time series based on one step ahead forecasts
}
\details{
Version 3.9, 5/24/2023
}
\examples{
air_terror <- r_terror(AirPassengers, log_transform = TRUE,      
                       aictest = c('td', 'easter'), terror_lags = 3) 
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
