% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.r
\name{replace_na}
\alias{replace_na}
\title{Replace NA}
\usage{
replace_na(this_vec = NULL, replace_string = "NA", replace_number = NULL)
}
\arguments{
\item{this_vec}{Vector object. This is a required entry.}

\item{replace_string}{Character scalar which replaces the NAs in the vector.
Default is \code{'NA'}.}

\item{replace_number}{Number which replaces the \code{NA}s in the vector.
Default is the \code{NA} is replaced by the string in \code{replace_string}.}
}
\value{
A vector with all \code{NA}s replaced by either a character string or a number.
}
\description{
Replace NA in a vector with a string or number.
}
\details{
Version 2.4, 5/25/2023
}
\examples{
sample_vec <- c(rnorm(25), NA, rnorm(24))
sample_vec_missing  <- replace_na(sample_vec, replace_string = 'Missing')
sample_vec_missing_number  <- replace_na(sample_vec, replace_number = -9999)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
