% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_seas_object.r
\name{save_seas_object}
\alias{save_seas_object}
\title{Save seas objects}
\usage{
save_seas_object(
  seas_obj = NULL,
  file_name = NULL,
  series_name = NULL,
  data_list = NULL,
  list_element = NULL,
  user_reg = NULL,
  this_window = FALSE,
  this_directory = NULL,
  this_sep = "_",
  print_out = FALSE
)
}
\arguments{
\item{seas_obj}{seasonal object. This is a required entry.}

\item{file_name}{character string; file name where \code{seas} object is stored;
default is the name of the seasonal object.}

\item{series_name}{character string; name of time series object used by the \code{seas} object;
default is the name of the seasonal object.}

\item{data_list}{character string; name of the list object that holds data; there is no default.}

\item{list_element}{character string; name of the list element used as data;
default is the name of the seasonal object.}

\item{user_reg}{character string; name of a time series matrix containing user defined
regressors; there is no default.
If not set, will set variables related to user defined regressors to \code{NULL}
in the static version of the \code{seas} object.}

\item{this_window}{logical indicator variable; determines if a span of the original series
If \code{FALSE}, the entire series will be used in the saved file.}

\item{this_directory}{character string; optional directory where the spec file is stored.}

\item{this_sep}{character string; separator between elements of the file name.
Default is \code{"_"}.}

\item{print_out}{logical indicator variable; determines if an \code{out()} function is
printed at the end of the script. If \code{FALSE}, the \code{out()} function is commented out.
Default is \code{FALSE}.}
}
\value{
stores the \code{seas} command to reproduce the \code{seas} object
\code{seas_obj} into the file \code{file_name.r} -
if \code{file_name} is not specified, the name of the seasonal object will be used
to form the output file name.
}
\description{
Stores \code{seas} command to reproduce the \code{seas} object \code{seas_obj}
into the file \code{file_name.r}.
}
\details{
Version 9.5, 5/14/2024
}
\examples{
ukgas_seas <-
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)",
                  x11 = "", transform.function = "log", forecast.maxlead = 20,
                  check.print = c( "pacf", "pacfplot" ))
\dontrun{save_seas_object(ukgas_seas, file_name = "ukgas_seas", series_name = "ukgas",
                 print_out = TRUE)}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
