% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_spec_file.r
\name{save_spec_file}
\alias{save_spec_file}
\title{Save spec file representation of \code{seas} object}
\usage{
save_spec_file(
  seas_obj = NULL,
  file_name = NULL,
  this_directory = NULL,
  this_data_directory = NULL,
  data_file_name = NULL,
  xreg_file_name = NULL,
  this_user_name = NULL,
  this_title = NULL
)
}
\arguments{
\item{seas_obj}{\code{seasonal} object. This is a required entry.}

\item{file_name}{character string; file name where \code{seas} object is stored;
default is the name of the \code{seasonal} object}

\item{this_directory}{character string; optional directory where the spec file is stored.}

\item{this_data_directory}{character string; optional directory where the data files are stored.
Default is no change in \code{file} entry in the spec file.}

\item{data_file_name}{character string; optional external file name where data file is stored.
Path should be included with file name if data file is not in working directory;
quotes will be added by the routine.
Default is no change in \code{file} entry in the spec file.}

\item{xreg_file_name}{character string; optional external file name where user defined
regressors are stored. Path should be included with file name if data file is not in
working directory; quotes will be added by the routine.
Default is no change in \code{file} entry in the spec file.}

\item{this_user_name}{vector of character strings; optional names for the user-defined
regressors. Should only appear if \code{xreg_file_name} is specified.}

\item{this_title}{character string; optional custom title; quotes will be added by the routine.
Default is no change in \code{title} entry in the spec file.}
}
\value{
stores the spec file representation of the \code{seas} object \code{seas_obj}
into the file \code{file_name.spc}
}
\description{
Stores the spec file representation of the \code{seas} object \code{seas_obj}
into the file \code{file_name.spc}.
}
\details{
Version 4.3, 5/14/2024
}
\examples{
ukgas_seas <- seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                             x11="", transform.function = "log", forecast.maxlead=20,
                             check.print = c( "pacf", "pacfplot" ))
\dontrun{save_spec_file(ukgas_seas, 'ukgas', 
                       data_file_name = "ukgas.dat",
                       this_directory = "X:\\\\seasonalAdj\\\\testing\\\\sautilities",
                       this_title = "Production run for quarterly UK Gas")}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
