% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_user_reg.r
\name{save_user_reg}
\alias{save_user_reg}
\title{save user regression matrix}
\usage{
save_user_reg(
  this_reg = NULL,
  this_reg_file = NULL,
  add_dates = TRUE,
  start_date = NULL
)
}
\arguments{
\item{this_reg}{double precision time series array or matrix to be saved.
This is a required argument.}

\item{this_reg_file}{character string; name of file time series array or matrix to be saved to.
This is a required argument.}

\item{add_dates}{logical scalar; save the file with dates on each record (datevalue format).
Default is TRUE.}

\item{start_date}{Integer vector of length 2, Start date for series to be stored, where the
first element is the year and the second element is the month or quarter..
Default is start of series.}
}
\value{
file with user-defined regressors will be produced.
}
\description{
Save a user-defined regression array or matrix with time series attributes to an external ASCII file
in X-13ARIMA-SEATS' datevalue format
}
\details{
Version 3.2, 12/27/2024
}
\examples{
UKgas_tc_date <- c(1970, 2)
UKgas_tc_1970_2 <-
   gen_tc_outlier_ts(UKgas_tc_date, this_start = c(1960, 1), this_end = c(1990, 4), 
                     this_freq = 4, tc_alpha = 0.9, return_matrix = TRUE)
\dontrun{save_user_reg(UKgas_tc_1970_2, 'UKgas_tc_1970_2.txt')}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
