% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_ftest.r
\name{seasonal_ftest}
\alias{seasonal_ftest}
\title{Model-based F-Test for Time Series Models.}
\usage{
seasonal_ftest(
  seas_obj = NULL,
  p_limit_fail = 0.01,
  p_limit_warn = 0.05,
  use_seasonal = TRUE,
  return_this = "test"
)
}
\arguments{
\item{seas_obj}{object generated by \code{seas()} of the seasonal package.
This is a required entry.}

\item{p_limit_fail}{Numeric scalar; P-value limit for model based seasonal F-statistic
for passing. Default is \code{0.01}.}

\item{p_limit_warn}{Numeric scalar; P-value limit for model based seasonal F-statistic
for a warning. Default is \code{0.05}.}

\item{use_seasonal}{logical scalar; if TRUE, the \code{seasonal} regressor is used;
otherwise, use the sine-cosine trignonmetric regressors generated by \code{sincos}.}

\item{return_this}{character string; what the function returns -
\code{'test'} returns test results, \code{'why'} returns why the test failed or
received a warning, or \code{'both'}. Default is \code{'test'}.}
}
\value{
A text string denoting if the series passed or failed tests for seasonality
using the model based F-test diagnostic.
}
\description{
Model based test for seasonality based on stable seasonal regressors
}
\details{
Version 5.2, 5/14/2024
}
\examples{
m_air <- 
  seasonal::seas(AirPassengers, transform.function = "log", arima.model = '(0 1 1)', 
                 regression.variables = c('seasonal', 'td'), x11="")
this_seasonal_ftest <- seasonal_ftest(m_air, return_this = 'both')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
