% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_critical_value.r
\name{set_critical_value}
\alias{set_critical_value}
\title{Set outlier critical value}
\usage{
set_critical_value(number_observations = NULL, cv_alpha = 0.01)
}
\arguments{
\item{number_observations}{number of observations tested for outliers
This is a required entry.}

\item{cv_alpha}{alpha for critical value}
}
\value{
outlier critical value generated by the algorithm given in Ljung (1993).
The critical value in X-13 is different as it is adjusted to allow for smaller values
to approximate the normal distribution.
}
\description{
Set outlier critical value using the Ljung algorithm as given in
Ljung, G. M. (1993). On outlier detection in time series.
Journal of Royal Statistical Society B 55, 559-567.
}
\details{
Version 1.6 5/25/2023
}
\examples{
this_critical_value <- set_critical_value(12, 0.025)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
