% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_legend_position.r
\name{set_legend_position}
\alias{set_legend_position}
\title{generate position of plot legend}
\usage{
set_legend_position(
  data_matrix = NULL,
  this_plot_start = NULL,
  this_plot_freq = 12,
  time_disp = 3,
  value_disp = 1/6,
  default_code = "top"
)
}
\arguments{
\item{data_matrix}{numeric matrix; matrix where all series being plotted are stored as columns.
This is a required entry.}

\item{this_plot_start}{Integer scalar; start date of the plot.
This is a required entry.}

\item{this_plot_freq}{Integer scalar; Frequency of time series plotted. Default is 12.}

\item{time_disp}{Integer scalar; number of observations on the x-axis taken up by the legend.
Default is 3.}

\item{value_disp}{Numeric scalar; factor representing the percentage of the y axis taken up
by the legend. Default is 1/6.}

\item{default_code}{Character string; default position code if the corners are not available.
Default is \code{"top"}. Possible values are \code{"bottomright"}, \code{"bottom"},
\code{"bottomleft"}, \code{"left"}, \code{"topleft"}, \code{"topright"},
\code{"top"}, \code{"right"} and \code{"center"}.}
}
\value{
Position codes for the legend command. Possible values are \code{"bottomright"},
\code{"bottom"}, \code{"bottomleft"}, \code{"topleft"}, \code{"topright"}
and the value of \code{default_code}.
}
\description{
Generate position code for the \code{legend} command based on the series being plotted.
}
\details{
Version 2.8, 5/25/2023
}
\examples{
shoes_seas <- 
    seasonal::seas(shoes2007, slidingspans = "", transform.function = "log", x11 = "",
                   forecast.maxlead=36, check.print = c( "pacf", "pacfplot" ))
this_series <- shoes2007
this_sa     <- seasonal::final(shoes_seas)
this_legend_position <- 
    set_legend_position(cbind(this_series, this_sa), start(this_series),
                        this_plot_freq = 4, time_disp = 8, value_disp = 1/8,
                        default_code = "top")                      
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
