% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoes2007.r
\docType{data}
\name{shoes2007}
\alias{shoes2007}
\title{Retail sales of shoes, 2007}
\format{
Retail sales of shoes ending in December of 2007
}
\usage{
shoes2007
}
\description{
A time series object
}
\keyword{datasets}
