% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significant_t_stats.r
\name{significant_t_stats}
\alias{significant_t_stats}
\title{Number of signifcant uregressors}
\usage{
significant_t_stats(seas_obj = NULL, critical_value = 2, user_def_only = FALSE)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas}
on a single time series. A required argument.}

\item{critical_value}{numeric scalar; sets the critical value used to test t-statistics of regressors.
Default is 2.0}

\item{user_def_only}{Logical scalar; only test user-defined regressors
Default is FALSE.}
}
\value{
number of regressors with significant t-statistics
}
\description{
Determines the number of regressors in a \code{seas} object that are significant.
}
\details{
Version 2.0, 5/1/2025
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, x11="", slidingspans = "", 
                  transform.function = "log", arima.model = "(0 1 1)(0 1 1)", 
                  regression.aictest = "td", forecast.maxlead=36, 
                  check.print = c( "pacf", "pacfplot" ), check.save = "acf")
air_num_user_reg <- significant_t_stats(air_seas)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
