% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_peak_fail_why.r
\name{spec_peak_fail_why}
\alias{spec_peak_fail_why}
\title{Failure text for spectral peaks}
\usage{
spec_peak_fail_why(
  udg_list = NULL,
  peak_level = 6,
  this_spec = "spcsa",
  return_both = FALSE
)
}
\arguments{
\item{udg_list}{\itemize{
\item list object generated by \code{udg()} function of the \code{seasonal} package.
This is a required entry.
}}

\item{peak_level}{Integer scalar - limit to determine if a frequency has a spectral peak.}

\item{this_spec}{text string with the spectrum being tested allowable entries are
\code{'spcori'}, \code{'spcsa'}, \code{'spcirr'}, \code{'spcrsd'}.
Default is \code{'spcori'}.}

\item{return_both}{Logical scalar indicating whether the calling function will return both
the test results and why the test failed or produced a warning. Default is \code{FALSE}.}
}
\value{
A text string denoting if the series passed the tests of spectrum diagnostics,
or why the series did not pass. Note that for \code{'spcori'}, the series fails
}
\description{
generate text on why spectral peaks are flagged
}
\details{
Version 3.1, 5/24/2023
}
\examples{
air_seas <- seasonal::seas(AirPassengers, transform.function = "log",  
                        arima.model = "(0 1 1)(0 1 1)", x11 = "")
air_seas_udg <- seasonal::udg(air_seas)
this_spec_peak_fail_why <- 
    spec_peak_fail_why(air_seas_udg, this_spec = 'spcori', return_both = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
