% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_peak_test.r
\name{spec_peak_test}
\alias{spec_peak_test}
\title{Test for spectral peaks}
\usage{
spec_peak_test(
  seas_obj = NULL,
  peak_level = 6,
  peak_warn = 3,
  this_spec = "spcsa",
  return_this = "test"
)
}
\arguments{
\item{seas_obj}{object generated by \code{seas()} of the \code{seasonal} package.
This is a required entry.}

\item{peak_level}{Integer scalar - limit to determine if a frequency has a spectral peak.
Default is 6.}

\item{peak_warn}{Integer scalar - limit to produce a warning that a frequency may have a
spectral peak. Default is 3.}

\item{this_spec}{text string with the spectrum being tested. Allowable entries are
\code{'spcori'}, \code{'spcsa'}, \code{'spcirr'}, \code{'spcrsd'}.
Default is \code{'spcori'}.}

\item{return_this}{character string; what the function returns -
\code{'test'} returns test results, \code{'why'} returns why the test failed or
received a warning, or \code{'both'}. Default is \code{'test'}.}
}
\value{
A text string denoting if the series passed or failed the tests of spectrum
diagnostics. Note that for \code{spcori}, the series fails if none of the
frequencies tested had peaks.
}
\description{
Test if spectral peaks are flagged.
}
\details{
Version 4.0, 05/3/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, transform.function = "log", 
                        arima.model = "(0 1 1)(0 1 1)", x11 = "")
this_spec_peak_test <- spec_peak_test(air_seas, this_spec = 'spcori', return_this = 'both')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
