% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_peak_warn_why.r
\name{spec_peak_warn_why}
\alias{spec_peak_warn_why}
\title{Warning message for spectral peaks}
\usage{
spec_peak_warn_why(
  udg_list = NULL,
  peak_warn_level = 3,
  this_spec = "spcsa",
  return_both = FALSE
)
}
\arguments{
\item{udg_list}{List object generated by \code{udg()} function of the \code{seasonal} package.
This is a required entry.}

\item{peak_warn_level}{Integer scalar - limit to produce a warning that a frequency may
have a spectral peak. Default is 3.}

\item{this_spec}{text string with the spectrum being tested. Allowable entries are
\code{'spcori'}, \code{'spcsa'}, \code{'spcirr'}, \code{'spcrsd'}.
Default is \code{'spcori'}.}

\item{return_both}{Logical scalar indicating whether the calling function will return both
the test results and why the test failed or produced a warning. Default is \code{FALSE}.}
}
\value{
A text string denoting if the series passed the tests of spectrum diagnostics,
or why the series did not pass. Note that for spcori, the series fails if none of
the frequencies tested had peaks
}
\description{
Generate warning message related to spectral peaks for a given X-13ARIMA-SEATS spectrum.
}
\details{
Version 3.2, 5/24/2023
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, transform.function = "log",  
                  arima.model = "(0 1 1)(0 1 1)", x11 = "")
air_seas_udg <- seasonal::udg(air_seas)
this_spec_peak_warn_why <- 
    spec_peak_warn_why(air_seas_udg, this_spec = 'spcori', return_both = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
