% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspan_test.r
\name{sspan_test}
\alias{sspan_test}
\title{Sliding Spans Diagnostic}
\usage{
sspan_test(
  seas_obj = NULL,
  sf_limit = 25,
  change_limit = 40,
  additivesa = FALSE,
  return_this = "test"
)
}
\arguments{
\item{seas_obj}{object generated by \code{seas()} of the seasonal package.
This is a required entry.}

\item{sf_limit}{Numeric object; limit for the percentage of seasonal spans flagged.
Default is 25.}

\item{change_limit}{Numeric object; limit for the percentage of month-to-month changes flagged.
Default is 40.}

\item{additivesa}{logical scalar; if true, the adjustment is assumed to be additive;
default is \code{FALSE}.}

\item{return_this}{character string; what the function returns -
\code{'test'} returns test results, \code{'why'} returns why the test failed or
received a warning, or \code{'both'}. Default is \code{'test'}.}
}
\value{
A text string denoting if the series passed or failed the tests of sliding spans diagnostics.
}
\description{
Results from using the sliding spans diagnostic.
}
\details{
Version 3.8, 5/14/2024
}
\examples{
ukgas_seas <- seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                             x11="", transform.function = "log", forecast.maxlead=20, 
                             slidingspans = "", check.print = c( "pacf", "pacfplot" ))
ukgas_sspan_test <- 
   sspan_test(ukgas_seas, sf_limit = 15, change_limit = 35, return_this = 'both')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
