% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspan_test_why.r
\name{sspan_test_why}
\alias{sspan_test_why}
\title{Sliding Spans Diagnostic Warning Messages}
\usage{
sspan_test_why(
  udg_list = NULL,
  sf_limit = 25,
  change_limit = 40,
  additivesa = FALSE,
  return_both = FALSE
)
}
\arguments{
\item{udg_list}{List object generated by \code{udg()} function of the \code{seasonal} package.
This is a required entry.}

\item{sf_limit}{Numeric object; limit for the percentage of seasonal spans flagged.
Default is 25.}

\item{change_limit}{Numeric object; limit for the percentage of month-to-month changes flagged.
Default is 40.}

\item{additivesa}{logical scalar; if true, the adjustment is assumed to be additive;
default is \code{FALSE}.}

\item{return_both}{Logical scalar indicating whether the calling function will return both
the test results and why the test failed or produced a warning. Default is \code{TRUE}.}
}
\value{
A text string denoting if the series passed the tests of sliding spans diagnostics
diagnostics, or why the series failed.
}
\description{
Generate text on why tests using the sliding spans diagnostic fail.
}
\details{
Version 4.2, 5/24/2023
}
\examples{
ukgas_seas <- seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                             x11="", transform.function = "log", forecast.maxlead=20,
                             check.print = c( "pacf", "pacfplot" ))
ukgas_seas_udg <- seasonal::udg(ukgas_seas)
ukgas_sspan_test_why <- 
   sspan_test_why(ukgas_seas_udg, sf_limit = 15, change_limit = 35, return_both = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
