% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_with_outlier_list.r
\name{static_with_outlier_list}
\alias{static_with_outlier_list}
\title{Add outliers to list of \code{seas} object}
\usage{
static_with_outlier_list(
  seas_obj_list = NULL,
  new_data_list = NULL,
  outlier_span = ",",
  outlier_types = "ao,ls"
)
}
\arguments{
\item{seas_obj_list}{list of seasonal objects. This is a required entry.}

\item{new_data_list}{list of time series objects; updated data sets from the data used
to generate \code{seas_obj_list}. This is a required entry.}

\item{outlier_span}{character string; sets the argument \code{outlier.span}.
Default is \code{","}.}

\item{outlier_types}{character string; sets the argument \code{outlier.types}.
Default is \code{"ao,ls"}.}
}
\value{
A list of updated static seas object with outlier arguments included.
}
\description{
Add outlier arguments to each element of a list of \code{seas} objects .
}
\details{
Version 3.4, 5/25/2023
}
\examples{
xt_seas_old <- 
   seasonal::seas(xt_data_old, slidingspans = "", 
                  transform.function = "log", 
                  x11 = "", forecast.maxlead = 60)
xt_outlier_seas <- 
   static_with_outlier_list(xt_seas_old, xt_data_new) 
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
