% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udg_series.r
\name{udg_series}
\alias{udg_series}
\title{Process a list of \code{seas} elements}
\usage{
udg_series(
  seas_obj_list = NULL,
  this_key = "autoout",
  this_limit = 5,
  this_abs = FALSE,
  greater_than = TRUE
)
}
\arguments{
\item{seas_obj_list}{\itemize{
\item list of \code{seas} objects generated by the \code{seasonal} package.
This is a required entry.
}}

\item{this_key}{\itemize{
\item character string containing keyword of the \code{udg} function
that returns a numeric value. Default is \code{'autoout'}.
}}

\item{this_limit}{\itemize{
\item numeric object which serves as the limit of the diagnostic
referred to in \code{this_key}. Default is \code{'5'}
}}

\item{this_abs}{Logical scalar that indicates whether the absolute value is taken
of the numbers before the comparison is made. (default is \code{FALSE})}

\item{greater_than}{\itemize{
\item logical object that specified whether the element names returned are
greater than or less than the limit specified in \code{this_limit}.
Default is \code{TRUE}.
}}
}
\value{
A vector of list element names where this_key is greater than or less than the limit
specified in \code{this_limit}. If nothing matches, the function will output the
string \code{'none'}.
}
\description{
Process a list of \code{seas} elements to find the elements that are greater than or
less than a particular limit for a diagnostic.
}
\details{
Version 3.3, 5/24/2023
}
\examples{
emp_seas_list <- seasonal::seas(employment_data_mts, 
                      slidingspans = "", transform.function = "log", 
                      x11 = "", arima.model = "(0 1 1)(0 1 1)",
                      forecast.maxlead=36, check.print = c( "pacf", "pacfplot" ))
emp_seas_update <- 
   Filter(function(x) inherits(x, "seas"), emp_seas_list)
xt_bad_m7 <- 
   udg_series(emp_seas_update, this_key = 'f3.m07', this_limit = 1.2)
xt_bad_q2 <- 
   udg_series(emp_seas_update, this_key = 'f3.qm2', this_limit = 1.2)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
