% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment_data_mts.r
\docType{data}
\name{unemployment_data_mts}
\alias{unemployment_data_mts}
\title{US Unemployment Series, four main components in an mts object}
\format{
An mts object with 4 time series elements in four columns:
\describe{
\item{n3000013}{Unemployed Males 16-19}
\item{n3000014}{Unemployed Females 16-19}
\item{n3000025}{Unemployed Males 20+}
\item{n3000026}{Unemployed Females 20+}
}
}
\usage{
unemployment_data_mts
}
\description{
#' An mts object of the four main components of US Unemployment expressed as time series objects that end in December, 2022
}
\keyword{datasets}
