% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment_list.r
\docType{data}
\name{unemployment_list}
\alias{unemployment_list}
\title{US Unemployment Series, four main components in a list object}
\format{
A list object with 4 time series elements:
\describe{
\item{n3000013}{Unemployed Males 16-19}
\item{n3000014}{Unemployed Females 16-19}
\item{n3000025}{Unemployed Males 20+}
\item{n3000026}{Unemployed Females 20+}
}
}
\usage{
unemployment_list
}
\description{
#' A list object of the four main components of US Unemployment expressed as time series objects that end in December, 2022
}
\keyword{datasets}
