% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_diag_matrix.r
\name{update_diag_matrix}
\alias{update_diag_matrix}
\title{Update Diagnostic Matrix}
\usage{
update_diag_matrix(
  this_diag_list = NULL,
  this_test_list = NULL,
  this_label = NULL
)
}
\arguments{
\item{this_diag_list}{list object with elements for seasonal adjustment or modeling diagnostic,
titles, and the number of columns. This is a required entry.}

\item{this_test_list}{list object of a specific seasonal adjustment or modeling diagnostic.
This is a required entry.}

\item{this_label}{character string; name of diagnostic in \code{this_test_list}.
This is a required entry.}
}
\value{
list object with updated elements for seasonal adjustment or modeling diagnostic,
titles, and the number of columns.
}
\description{
Update the matrix of diagnostics used to generate the diagnostic data frame
in \code{make_diag_df}.
}
\details{
Version 4.3, 5/24/2023
}
\examples{
unemp_seas_list <- seasonal::seas(unemployment_list,
           x11 = "", slidingspans = "",
           arima.model = "(0 1 1)(0 1 1)",
           transform.function = "log",
           regression.aictest = NULL, 
           forecast.maxlead=60,
           check.print = c( "pacf", "pacfplot" ))
test_seas_update <- 
     Filter(function(x) inherits(x, "seas"), unemp_seas_list)
test_acf <- lapply(test_seas_update, function(x) try(acf_test(x, return_this = 'both')))
test_names <- names(xt_data_new)
num_names <- length(test_names)
all_diag_list <- list(n = 0, diag = 0, titles = 0)
if (!is.null(test_acf)) {
    if (length(test_acf) < num_names) {
        this_acf_test <- fix_diag_list(test_acf, test_names, return_this = 'both')
    }
    all_diag_list <- 
        update_diag_matrix(all_diag_list, test_acf, "ACF")
}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
