% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_error.r
\name{which_error}
\alias{which_error}
\title{Check list for try errors}
\usage{
which_error(this_list = NULL)
}
\arguments{
\item{this_list}{list object which potentially contains \code{'try-error'} class objects.}
}
\value{
vector of the names of list elements that are \code{'try-error'} class objects.
If the list contains no \code{'try-error'} class objects, the function will
return \code{NULL}.
}
\description{
Checks list for try errors, returning element names with errors
}
\details{
Version 2.9, 5/4/2024
}
\examples{
unemp_seas_list <- 
  seasonal::seas(unemployment_list, slidingspans = "", 
                 transform.function = "log", forecast.maxlead = 36, 
                 arima.model = "(0 1 1)(0 1 1)",
                 check.print = c( "pacf", "pacfplot" ))
unemp_seas_update <- 
     Filter(function(x) inherits(x, "seas"), unemp_seas_list)
unemp_seas_errors <- which_error(unemp_seas_update) 
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
