% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt_data_list.r
\docType{data}
\name{xt_data_list}
\alias{xt_data_list}
\title{US Building Permits}
\format{
A list object with 12 time series elements:
\describe{
\item{mw1u}{Midwest one family building permits}
\item{mwto}{Midwest total building permits}
\item{ne1u}{Northeast one family building permits}
\item{neto}{Northeast total building permits}
\item{so1u}{South one family building permits}
\item{soto}{South total building permits}
\item{we1u}{West one family building permits}
\item{weto}{West total building permits}
\item{us1u}{US one family building permits}
\item{us24}{US 2-4 family building permits}
\item{us5p}{US 5+ family building permits}
\item{usto}{US total family building permits}
}
}
\usage{
xt_data_list
}
\description{
A list object with 12 components of US Building Permits expressed as time series objects
}
\keyword{datasets}
